@@ -1,16 +1,33 @@
-PROJECT(local_sensing_node)
-CMAKE_MINIMUM_REQUIRED(VERSION 2.8.3)
-SET(CMAKE_BUILD_TYPE Release) # Release, RelWithDebInfo
-#SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -std=c++11")
-ADD_COMPILE_OPTIONS(-std=c++11 )
-ADD_COMPILE_OPTIONS(-std=c++14 )
+cmake_minimum_required(VERSION 3.8)
+project(local_sensing)
 
-#set(ENABLE_CUDA false)
+# 设置编译选项
+set(CMAKE_BUILD_TYPE Release)
+add_compile_options(-std=c++17)
+
+if(CMAKE_COMPILER_IS_GNUCXX OR CMAKE_CXX_COMPILER_ID MATCHES "Clang")
+  add_compile_options(-Wall -Wextra -Wpedantic)
+endif()
+
 set(ENABLE_CUDA false)
+# set(ENABLE_CUDA true)
+
+
+if(BUILD_TESTING)
+  find_package(ament_lint_auto REQUIRED)
+  # the following line skips the linter which checks for copyrights
+  # comment the line when a copyright and license is added to all source files
+  set(ament_cmake_copyright_FOUND TRUE)
+  # the following line skips cpplint (only works in a git repo)
+  # comment the line when this package is in a git repo and when
+  # a copyright and license is added to all source files
+  set(ament_cmake_cpplint_FOUND TRUE)
+  ament_lint_auto_find_test_dependencies()
+endif()
 
 if(ENABLE_CUDA)
   find_package(CUDA REQUIRED)
-  SET(CUDA_NVCC_FLAGS ${CUDA_NVCC_FLAGS};-O3 -use_fast_math)
+  set(CUDA_NVCC_FLAGS ${CUDA_NVCC_FLAGS};-O3 -use_fast_math)
   set(CUDA_NVCC_FLAGS 
 #       -gencode arch=compute_20,code=sm_20;
 #       -gencode arch=compute_20,code=sm_21;
