@@ -1,59 +1,102 @@
 #include "plan_env/grid_map.h"
+#include <limits>
+#include <tf2_ros/buffer.h>
+#include <tf2_ros/transform_listener.h>
+#include <tf2_geometry_msgs/tf2_geometry_msgs.hpp>
+#include <tf2_eigen/tf2_eigen.hpp>
+#include <pcl/common/transforms.h>
 
 // #define current_img_ md_.depth_image_[image_cnt_ & 1]
 // #define last_img_ md_.depth_image_[!(image_cnt_ & 1)]
 
-void GridMap::initMap(ros::NodeHandle &nh)
+void GridMap::initMap(rclcpp::Node::SharedPtr node)
 {
-  node_ = nh;
+  node_ = node;
+
+  // 初始化TF2缓冲区（全局缓存，避免重复创建）
+  if (!tf_buffer_) {
+    tf_buffer_ = std::make_shared<tf2_ros::Buffer>(node_->get_clock());
+    tf_listener_ = std::make_shared<tf2_ros::TransformListener>(*tf_buffer_);
+    RCLCPP_INFO(node_->get_logger(), "TF2缓冲区初始化完成");
+  }
 
   /* get parameter */
   double x_size, y_size, z_size;
-  node_.param("grid_map/resolution", mp_.resolution_, -1.0);
-  node_.param("grid_map/map_size_x", x_size, -1.0);
-  node_.param("grid_map/map_size_y", y_size, -1.0);
-  node_.param("grid_map/map_size_z", z_size, -1.0);
-  node_.param("grid_map/local_update_range_x", mp_.local_update_range_(0), -1.0);
-  node_.param("grid_map/local_update_range_y", mp_.local_update_range_(1), -1.0);
-  node_.param("grid_map/local_update_range_z", mp_.local_update_range_(2), -1.0);
-  node_.param("grid_map/obstacles_inflation", mp_.obstacles_inflation_, -1.0);
+  node_->declare_parameter("grid_map/resolution", -1.0);
+  node_->declare_parameter("grid_map/map_size_x", -1.0);
+  node_->declare_parameter("grid_map/map_size_y", -1.0);
+  node_->declare_parameter("grid_map/map_size_z", -1.0);
+  node_->declare_parameter("grid_map/local_update_range_x", -1.0);
+  node_->declare_parameter("grid_map/local_update_range_y", -1.0);
+  node_->declare_parameter("grid_map/local_update_range_z", -1.0);
+  node_->declare_parameter("grid_map/obstacles_inflation", -1.0);
+  node_->declare_parameter("grid_map/fx", -1.0);
+  node_->declare_parameter("grid_map/fy", -1.0);
+  node_->declare_parameter("grid_map/cx", -1.0);
+  node_->declare_parameter("grid_map/cy", -1.0);
+  node_->declare_parameter("grid_map/use_depth_filter", true);
+  node_->declare_parameter("grid_map/depth_filter_tolerance", -1.0);
+  node_->declare_parameter("grid_map/depth_filter_maxdist", -1.0);
+  node_->declare_parameter("grid_map/depth_filter_mindist", -1.0);
+  node_->declare_parameter("grid_map/depth_filter_margin", -1);
+  node_->declare_parameter("grid_map/k_depth_scaling_factor", -1.0);
+  node_->declare_parameter("grid_map/skip_pixel", -1);
+  node_->declare_parameter("grid_map/p_hit", 0.70);
+  node_->declare_parameter("grid_map/p_miss", 0.35);
+  node_->declare_parameter("grid_map/p_min", 0.12);
+  node_->declare_parameter("grid_map/p_max", 0.97);
+  node_->declare_parameter("grid_map/p_occ", 0.80);
+  node_->declare_parameter("grid_map/min_ray_length", -0.1);
+  node_->declare_parameter("grid_map/max_ray_length", -0.1);
+  node_->declare_parameter("grid_map/visualization_truncate_height", -0.1);
+  node_->declare_parameter("grid_map/virtual_ceil_height", -0.1);
+  node_->declare_parameter("grid_map/virtual_ceil_yp", -0.1);
+  node_->declare_parameter("grid_map/virtual_ceil_yn", -0.1);
+  node_->declare_parameter("grid_map/show_occ_time", false);
+  node_->declare_parameter("grid_map/pose_type", 1);
+  node_->declare_parameter("grid_map/frame_id", "world");
+  node_->declare_parameter("grid_map/local_map_margin", 1);
+  node_->declare_parameter("grid_map/ground_height", 1.0);
+  node_->declare_parameter("grid_map/odom_depth_timeout", 1.0);
 
-  node_.param("grid_map/fx", mp_.fx_, -1.0);
-  node_.param("grid_map/fy", mp_.fy_, -1.0);
-  node_.param("grid_map/cx", mp_.cx_, -1.0);
-  node_.param("grid_map/cy", mp_.cy_, -1.0);
+  node_->get_parameter("grid_map/resolution", mp_.resolution_);
+  node_->get_parameter("grid_map/map_size_x", x_size);
+  node_->get_parameter("grid_map/map_size_y", y_size);
+  node_->get_parameter("grid_map/map_size_z", z_size);
+  node_->get_parameter("grid_map/local_update_range_x", mp_.local_update_range_(0));
+  node_->get_parameter("grid_map/local_update_range_y", mp_.local_update_range_(1));
+  node_->get_parameter("grid_map/local_update_range_z", mp_.local_update_range_(2));
+  node_->get_parameter("grid_map/obstacles_inflation", mp_.obstacles_inflation_);
+  node_->get_parameter("grid_map/fx", mp_.fx_);
+  node_->get_parameter("grid_map/fy", mp_.fy_);
+  node_->get_parameter("grid_map/cx", mp_.cx_);
+  node_->get_parameter("grid_map/cy", mp_.cy_);
+  node_->get_parameter("grid_map/use_depth_filter", mp_.use_depth_filter_);
+  node_->get_parameter("grid_map/depth_filter_tolerance", mp_.depth_filter_tolerance_);
+  node_->get_parameter("grid_map/depth_filter_maxdist", mp_.depth_filter_maxdist_);
+  node_->get_parameter("grid_map/depth_filter_mindist", mp_.depth_filter_mindist_);
+  node_->get_parameter("grid_map/depth_filter_margin", mp_.depth_filter_margin_);
+  node_->get_parameter("grid_map/k_depth_scaling_factor", mp_.k_depth_scaling_factor_);
+  node_->get_parameter("grid_map/skip_pixel", mp_.skip_pixel_);
+  node_->get_parameter("grid_map/p_hit", mp_.p_hit_);
+  node_->get_parameter("grid_map/p_miss", mp_.p_miss_);
+  node_->get_parameter("grid_map/p_min", mp_.p_min_);
+  node_->get_parameter("grid_map/p_max", mp_.p_max_);
+  node_->get_parameter("grid_map/p_occ", mp_.p_occ_);
+  node_->get_parameter("grid_map/min_ray_length", mp_.min_ray_length_);
+  node_->get_parameter("grid_map/max_ray_length", mp_.max_ray_length_);
+  node_->get_parameter("grid_map/visualization_truncate_height", mp_.visualization_truncate_height_);
+  node_->get_parameter("grid_map/virtual_ceil_height", mp_.virtual_ceil_height_);
+  node_->get_parameter("grid_map/virtual_ceil_yp", mp_.virtual_ceil_yp_);
+  node_->get_parameter("grid_map/virtual_ceil_yn", mp_.virtual_ceil_yn_);
+  node_->get_parameter("grid_map/show_occ_time", mp_.show_occ_time_);
+  node_->get_parameter("grid_map/pose_type", mp_.pose_type_);
+  node_->get_parameter("grid_map/frame_id", mp_.frame_id_);
+  node_->get_parameter("grid_map/local_map_margin", mp_.local_map_margin_);
+  node_->get_parameter("grid_map/ground_height", mp_.ground_height_);
+  node_->get_parameter("grid_map/odom_depth_timeout", mp_.odom_depth_timeout_);
 
-  node_.param("grid_map/use_depth_filter", mp_.use_depth_filter_, true);
-  node_.param("grid_map/depth_filter_tolerance", mp_.depth_filter_tolerance_, -1.0);
-  node_.param("grid_map/depth_filter_maxdist", mp_.depth_filter_maxdist_, -1.0);
-  node_.param("grid_map/depth_filter_mindist", mp_.depth_filter_mindist_, -1.0);
-  node_.param("grid_map/depth_filter_margin", mp_.depth_filter_margin_, -1);
-  node_.param("grid_map/k_depth_scaling_factor", mp_.k_depth_scaling_factor_, -1.0);
-  node_.param("grid_map/skip_pixel", mp_.skip_pixel_, -1);
-
-  node_.param("grid_map/p_hit", mp_.p_hit_, 0.70);
-  node_.param("grid_map/p_miss", mp_.p_miss_, 0.35);
-  node_.param("grid_map/p_min", mp_.p_min_, 0.12);
-  node_.param("grid_map/p_max", mp_.p_max_, 0.97);
-  node_.param("grid_map/p_occ", mp_.p_occ_, 0.80);
-  node_.param("grid_map/min_ray_length", mp_.min_ray_length_, -0.1);
-  node_.param("grid_map/max_ray_length", mp_.max_ray_length_, -0.1);
-
-  node_.param("grid_map/visualization_truncate_height", mp_.visualization_truncate_height_, -0.1);
-  node_.param("grid_map/virtual_ceil_height", mp_.virtual_ceil_height_, -0.1);
-  node_.param("grid_map/virtual_ceil_yp", mp_.virtual_ceil_yp_, -0.1);
-  node_.param("grid_map/virtual_ceil_yn", mp_.virtual_ceil_yn_, -0.1);
-
-  node_.param("grid_map/show_occ_time", mp_.show_occ_time_, false);
-  node_.param("grid_map/pose_type", mp_.pose_type_, 1);
-
-  node_.param("grid_map/frame_id", mp_.frame_id_, string("world"));
-  node_.param("grid_map/local_map_margin", mp_.local_map_margin_, 1);
-  node_.param("grid_map/ground_height", mp_.ground_height_, 1.0);
-
-  node_.param("grid_map/odom_depth_timeout", mp_.odom_depth_timeout_, 1.0);
-
-  if( mp_.virtual_ceil_height_ - mp_.ground_height_ > z_size)
+  if (mp_.virtual_ceil_height_ - mp_.ground_height_ > z_size)
   {
     mp_.virtual_ceil_height_ = mp_.ground_height_ + z_size;
   }
@@ -105,39 +148,54 @@ void GridMap::initMap(ros::NodeHandle &nh)
 
   /* init callback */
 
-  depth_sub_.reset(new message_filters::Subscriber<sensor_msgs::Image>(node_, "grid_map/depth", 50));
-  extrinsic_sub_ = node_.subscribe<nav_msgs::Odometry>(
-      "/vins_estimator/extrinsic", 10, &GridMap::extrinsicCallback, this); //sub
+  // 初始化 message_filters::Subscriber
+  depth_sub_ = std::make_shared<message_filters::Subscriber<sensor_msgs::msg::Image>>(
+      node_, "grid_map/depth", rclcpp::QoS(50).get_rmw_qos_profile());
+
+  extrinsic_sub_ = node_->create_subscription<nav_msgs::msg::Odometry>(
+      "/vins_estimator/extrinsic", 10,
+      std::bind(&GridMap::extrinsicCallback, this, std::placeholders::_1));
 
   if (mp_.pose_type_ == POSE_STAMPED)
   {
-    pose_sub_.reset(
-        new message_filters::Subscriber<geometry_msgs::PoseStamped>(node_, "grid_map/pose", 25));
+    pose_sub_ = std::make_shared<message_filters::Subscriber<geometry_msgs::msg::PoseStamped>>(
+        node_, "grid_map/pose", rclcpp::QoS(25).get_rmw_qos_profile());
 
-    sync_image_pose_.reset(new message_filters::Synchronizer<SyncPolicyImagePose>(
-        SyncPolicyImagePose(100), *depth_sub_, *pose_sub_));
-    sync_image_pose_->registerCallback(boost::bind(&GridMap::depthPoseCallback, this, _1, _2));
+    sync_image_pose_ = std::make_shared<message_filters::Synchronizer<SyncPolicyImagePose>>(
+        SyncPolicyImagePose(100), *depth_sub_, *pose_sub_);
+    sync_image_pose_->registerCallback(
+        std::bind(&GridMap::depthPoseCallback, this, std::placeholders::_1, std::placeholders::_2));
   }
   else if (mp_.pose_type_ == ODOMETRY)
   {
-    odom_sub_.reset(new message_filters::Subscriber<nav_msgs::Odometry>(node_, "grid_map/odom", 100, ros::TransportHints().tcpNoDelay()));
+    odom_sub_ = std::make_shared<message_filters::Subscriber<nav_msgs::msg::Odometry>>(
+        node_, "grid_map/odom", rclcpp::QoS(100).get_rmw_qos_profile());
 
-    sync_image_odom_.reset(new message_filters::Synchronizer<SyncPolicyImageOdom>(
-        SyncPolicyImageOdom(100), *depth_sub_, *odom_sub_));
-    sync_image_odom_->registerCallback(boost::bind(&GridMap::depthOdomCallback, this, _1, _2));
+    sync_image_odom_ = std::make_shared<message_filters::Synchronizer<SyncPolicyImageOdom>>(
+        SyncPolicyImageOdom(100), *depth_sub_, *odom_sub_);
+    sync_image_odom_->registerCallback(
+        std::bind(&GridMap::depthOdomCallback, this, std::placeholders::_1, std::placeholders::_2));
   }
 
-  // use odometry and point cloud
-  indep_cloud_sub_ =
-      node_.subscribe<sensor_msgs::PointCloud2>("grid_map/cloud", 10, &GridMap::cloudCallback, this);
-  indep_odom_sub_ =
-      node_.subscribe<nav_msgs::Odometry>("grid_map/odom", 10, &GridMap::odomCallback, this);
+  // 使用独立的里程计和点云订阅
+  indep_cloud_sub_ = node_->create_subscription<sensor_msgs::msg::PointCloud2>(
+      "grid_map/cloud", 1, std::bind(&GridMap::cloudCallback, this, std::placeholders::_1));
 
-  occ_timer_ = node_.createTimer(ros::Duration(0.05), &GridMap::updateOccupancyCallback, this);
-  vis_timer_ = node_.createTimer(ros::Duration(0.11), &GridMap::visCallback, this);
+  indep_odom_sub_ = node_->create_subscription<nav_msgs::msg::Odometry>(
+      "grid_map/odom", 1, std::bind(&GridMap::odomCallback, this, std::placeholders::_1));
 
-  map_pub_ = node_.advertise<sensor_msgs::PointCloud2>("grid_map/occupancy", 10);
-  map_inf_pub_ = node_.advertise<sensor_msgs::PointCloud2>("grid_map/occupancy_inflate", 10);
+  // 定时器
+  occ_timer_ = node_->create_wall_timer(
+      std::chrono::duration<double>(0.05),
+      std::bind(&GridMap::updateOccupancyCallback, this));
+
+  vis_timer_ = node_->create_wall_timer(
+      std::chrono::duration<double>(0.11),
+      std::bind(&GridMap::visCallback, this));
+
+  // 发布者
+  map_pub_ = node_->create_publisher<sensor_msgs::msg::PointCloud2>("grid_map/occupancy", 10);
+  map_inf_pub_ = node_->create_publisher<sensor_msgs::msg::PointCloud2>("grid_map/occupancy_inflate", 10);
 
   md_.occ_need_update_ = false;
   md_.local_updated_ = false;
@@ -145,7 +203,7 @@ void GridMap::initMap(ros::NodeHandle &nh)
   md_.has_odom_ = false;
   md_.has_cloud_ = false;
   md_.image_cnt_ = 0;
-  md_.last_occ_update_time_.fromSec(0);
+  md_.last_occ_update_time_ = rclcpp::Time(0, 0, RCL_SYSTEM_TIME);
 
   md_.fuse_time_ = 0.0;
   md_.update_num_ = 0;
@@ -229,11 +287,11 @@ void GridMap::projectDepthImage()
 
   if (!mp_.use_depth_filter_)
   {
-    for (int v = 0; v < rows; v+=skip_pix)
+    for (int v = 0; v < rows; v += skip_pix)
     {
       row_ptr = md_.depth_image_.ptr<uint16_t>(v);
 
-      for (int u = 0; u < cols; u+=skip_pix)
+      for (int u = 0; u < cols; u += skip_pix)
       {
 
         Eigen::Vector3d proj_pt;
@@ -342,7 +400,7 @@ void GridMap::raycastProcess()
   if (md_.proj_points_cnt == 0)
     return;
 
-  ros::Time t1, t2;
+  rclcpp::Time t1, t2;
 
   md_.raycast_num_ += 1;
 
@@ -644,37 +702,43 @@ void GridMap::clearAndInflateLocalMap()
       }
 
   // add virtual ceiling to limit flight height
-  if (mp_.virtual_ceil_height_ > -0.5) {
+  if (mp_.virtual_ceil_height_ > -0.05)
+  {
     int ceil_id = floor((mp_.virtual_ceil_height_ - mp_.map_origin_(2)) * mp_.resolution_inv_) - 1;
     for (int x = md_.local_bound_min_(0); x <= md_.local_bound_max_(0); ++x)
-      for (int y = md_.local_bound_min_(1); y <= md_.local_bound_max_(1); ++y) {
+      for (int y = md_.local_bound_min_(1); y <= md_.local_bound_max_(1); ++y)
+      {
         md_.occupancy_buffer_inflate_[toAddress(x, y, ceil_id)] = 1;
       }
   }
 }
 
-void GridMap::visCallback(const ros::TimerEvent & /*event*/)
+void GridMap::visCallback()
 {
-
   publishMapInflate(true);
   publishMap();
 }
 
-void GridMap::updateOccupancyCallback(const ros::TimerEvent & /*event*/)
+void GridMap::updateOccupancyCallback()
 {
-  if (md_.last_occ_update_time_.toSec() < 1.0 ) md_.last_occ_update_time_ = ros::Time::now();
-  
+  if (md_.last_occ_update_time_.seconds() < 1.0)
+    md_.last_occ_update_time_ = node_->now();
+
   if (!md_.occ_need_update_)
   {
-    if ( md_.flag_use_depth_fusion && (ros::Time::now() - md_.last_occ_update_time_).toSec() > mp_.odom_depth_timeout_ )
+    if (md_.flag_use_depth_fusion &&
+        (node_->now() - md_.last_occ_update_time_).seconds() > mp_.odom_depth_timeout_)
     {
-      ROS_ERROR("odom or depth lost! ros::Time::now()=%f, md_.last_occ_update_time_=%f, mp_.odom_depth_timeout_=%f", 
-        ros::Time::now().toSec(), md_.last_occ_update_time_.toSec(), mp_.odom_depth_timeout_);
+      RCLCPP_ERROR(node_->get_logger(),
+                   "odom or depth lost! now=%f, last_occ_update_time=%f, odom_depth_timeout=%f",
+                   node_->now().seconds(),
+                   md_.last_occ_update_time_.seconds(),
+                   mp_.odom_depth_timeout_);
       md_.flag_depth_odom_timeout_ = true;
     }
     return;
   }
-  md_.last_occ_update_time_ = ros::Time::now();
+  md_.last_occ_update_time_ = node_->now();
 
   /* update occupancy */
   // ros::Time t1, t2, t3, t4;
@@ -704,8 +768,8 @@ void GridMap::updateOccupancyCallback(const ros::TimerEvent & /*event*/)
   md_.local_updated_ = false;
 }
 
-void GridMap::depthPoseCallback(const sensor_msgs::ImageConstPtr &img,
-                                const geometry_msgs::PoseStampedConstPtr &pose)
+void GridMap::depthPoseCallback(const sensor_msgs::msg::Image::ConstPtr &img,
+                                const geometry_msgs::msg::PoseStamped::ConstPtr &pose)
 {
   /* get depth image */
   cv_bridge::CvImagePtr cv_ptr;
@@ -740,7 +804,7 @@ void GridMap::depthPoseCallback(const sensor_msgs::ImageConstPtr &img,
   md_.flag_use_depth_fusion = true;
 }
 
-void GridMap::odomCallback(const nav_msgs::OdometryConstPtr &odom)
+void GridMap::odomCallback(const nav_msgs::msg::Odometry::SharedPtr odom)
 {
   if (md_.has_first_depth_)
     return;
@@ -752,26 +816,90 @@ void GridMap::odomCallback(const nav_msgs::OdometryConstPtr &odom)
   md_.has_odom_ = true;
 }
 
-void GridMap::cloudCallback(const sensor_msgs::PointCloud2ConstPtr &img)
+void GridMap::cloudCallback(const sensor_msgs::msg::PointCloud2::ConstPtr &img)
 {
-
   pcl::PointCloud<pcl::PointXYZ> latest_cloud;
-  pcl::fromROSMsg(*img, latest_cloud);
+  
+  // 获取点云的时间戳
+  auto cloud_time = img->header.stamp;
+  
+  try
+  {
+    pcl::fromROSMsg(*img, latest_cloud);
+  }
+  catch (const std::exception& e)
+  {
+    RCLCPP_ERROR(node_->get_logger(), "点云数据转换失败: %s", e.what());
+    return;
+  }
 
   md_.has_cloud_ = true;
 
   if (!md_.has_odom_)
   {
-    std::cout << "no odom!" << std::endl;
+    RCLCPP_WARN(node_->get_logger(), "点云数据接收但odom数据缺失");
     return;
   }
 
   if (latest_cloud.points.size() == 0)
+  {
+    RCLCPP_WARN(node_->get_logger(), "点云数据点数为0");
     return;
+  }
 
   if (isnan(md_.camera_pos_(0)) || isnan(md_.camera_pos_(1)) || isnan(md_.camera_pos_(2)))
+  {
+    RCLCPP_WARN(node_->get_logger(), "相机位置数据无效");
     return;
-
+  }
+  
+  // 将点云从相机坐标系转换到世界坐标系
+  pcl::PointCloud<pcl::PointXYZ> world_cloud;
+  
+  try
+  {
+    // 使用全局TF2缓冲区（避免重复创建）
+    if (!tf_buffer_) {
+      RCLCPP_ERROR(node_->get_logger(), "TF2缓冲区未初始化，无法进行坐标系转换");
+      return;
+    }
+    
+    // 获取从相机坐标系到世界坐标系的变换
+    rclcpp::Duration timeout(-500000000, 500000000); // 500毫秒超时
+    geometry_msgs::msg::TransformStamped transform;
+    
+    // 根据点云头信息中的frame_id确定源坐标系
+    std::string source_frame = img->header.frame_id;
+    if (source_frame.empty())
+    {
+      source_frame = "x500_depth_0/StereoOV7251"; // 默认相机坐标系
+    }
+    
+    transform = tf_buffer_->lookupTransform(mp_.frame_id_, source_frame, cloud_time, timeout);
+    
+    // 获取变换矩阵
+    Eigen::Isometry3d transform_matrix = tf2::transformToEigen(transform);
+    
+    // 预分配内存避免重复分配
+    world_cloud.points.resize(latest_cloud.points.size());
+    
+    // 使用PCL的高效转换函数（需要将Eigen::Isometry3d转换为Eigen::Affine3d）
+    Eigen::Affine3d affine_transform(transform_matrix);
+    pcl::transformPointCloud(latest_cloud, world_cloud, affine_transform);
+    
+    RCLCPP_DEBUG(node_->get_logger(), "成功将点云从%s转换到world坐标系，点数: %zu", 
+                 source_frame.c_str(), world_cloud.points.size());
+  }
+  catch (tf2::TransformException &ex)
+  {
+    RCLCPP_ERROR(node_->get_logger(), "TF变换失败: %s，无法处理点云数据", ex.what());
+    return;
+  }
+  
+  // 使用转换后的世界坐标系点云
+  latest_cloud = world_cloud;
+  
+  // 在世界坐标系中重置缓冲区
   this->resetBuffer(md_.camera_pos_ - mp_.local_update_range_,
                     md_.camera_pos_ + mp_.local_update_range_);
 
@@ -805,6 +933,7 @@ void GridMap::cloudCallback(const sensor_msgs::PointCloud2ConstPtr &img)
     {
 
       /* inflate the point */
+      // 点云膨胀
       for (int x = -inf_step; x <= inf_step; ++x)
         for (int y = -inf_step; y <= inf_step; ++y)
           for (int z = -inf_step_z; z <= inf_step_z; ++z)
@@ -847,12 +976,15 @@ void GridMap::cloudCallback(const sensor_msgs::PointCloud2ConstPtr &img)
   posToIndex(Eigen::Vector3d(max_x, max_y, max_z), md_.local_bound_max_);
   posToIndex(Eigen::Vector3d(min_x, min_y, min_z), md_.local_bound_min_);
 
+  // 更新局部地图边界
   boundIndex(md_.local_bound_min_);
   boundIndex(md_.local_bound_max_);
 
   // add virtual ceiling to limit flight height
+  // 添加虚拟天花板控制飞行高度
   if (mp_.virtual_ceil_height_ > -0.5) {
     int ceil_id = floor((mp_.virtual_ceil_height_ - mp_.map_origin_(2)) * mp_.resolution_inv_) - 1;
+    // RCLCPP_INFO(node_->get_logger(), "添加虚拟天花板: 高度=%.2f, 栅格ID=%d", mp_.virtual_ceil_height_, ceil_id);
     for (int x = md_.local_bound_min_(0); x <= md_.local_bound_max_(0); ++x)
       for (int y = md_.local_bound_min_(1); y <= md_.local_bound_max_(1); ++y) {
         md_.occupancy_buffer_inflate_[toAddress(x, y, ceil_id)] = 1;
@@ -862,9 +994,10 @@ void GridMap::cloudCallback(const sensor_msgs::PointCloud2ConstPtr &img)
 
 void GridMap::publishMap()
 {
-
-  if (map_pub_.getNumSubscribers() <= 0)
+  if (map_pub_->get_subscription_count() <= 0)
+  {
     return;
+  }
 
   pcl::PointXYZ pt;
   pcl::PointCloud<pcl::PointXYZ> cloud;
@@ -900,17 +1033,15 @@ void GridMap::publishMap()
   cloud.width = cloud.points.size();
   cloud.height = 1;
   cloud.is_dense = true;
-  cloud.header.frame_id = mp_.frame_id_;
-  sensor_msgs::PointCloud2 cloud_msg;
-
+  cloud.header.frame_id = mp_.frame_id_;  // 点云已经在世界坐标系中
+  sensor_msgs::msg::PointCloud2 cloud_msg;
   pcl::toROSMsg(cloud, cloud_msg);
-  map_pub_.publish(cloud_msg);
+  map_pub_->publish(cloud_msg);
 }
 
-void GridMap::publishMapInflate(bool all_info)
-{
+void GridMap::publishMapInflate(bool all_info){
 
-  if (map_inf_pub_.getNumSubscribers() <= 0)
+  if (map_inf_pub_->get_subscription_count()<= 0)
     return;
 
   pcl::PointXYZ pt;
@@ -951,12 +1082,12 @@ void GridMap::publishMapInflate(bool all_info)
   cloud.height = 1;
   cloud.is_dense = true;
   cloud.header.frame_id = mp_.frame_id_;
-  sensor_msgs::PointCloud2 cloud_msg;
+  sensor_msgs::msg::PointCloud2 cloud_msg;
 
   pcl::toROSMsg(cloud, cloud_msg);
-  map_inf_pub_.publish(cloud_msg);
+  map_inf_pub_->publish(cloud_msg);
 
-  // ROS_INFO("pub map");
+  // RCLCPP_INFO(rclcpp::get_logger("publishMapInflate"), "pub map");
 }
 
 bool GridMap::odomValid() { return md_.has_odom_; }
@@ -974,7 +1105,7 @@ void GridMap::getRegion(Eigen::Vector3d &ori, Eigen::Vector3d &size)
   ori = mp_.map_origin_, size = mp_.map_size_;
 }
 
-void GridMap::extrinsicCallback(const nav_msgs::OdometryConstPtr &odom)
+void GridMap::extrinsicCallback(const nav_msgs::msg::Odometry::ConstPtr &odom)
 {
   Eigen::Quaterniond cam2body_q = Eigen::Quaterniond(odom->pose.pose.orientation.w,
                                                      odom->pose.pose.orientation.x,
@@ -988,8 +1119,8 @@ void GridMap::extrinsicCallback(const nav_msgs::OdometryConstPtr &odom)
   md_.cam2body_(3, 3) = 1.0;
 }
 
-void GridMap::depthOdomCallback(const sensor_msgs::ImageConstPtr &img,
-                                const nav_msgs::OdometryConstPtr &odom)
+void GridMap::depthOdomCallback(const sensor_msgs::msg::Image::ConstPtr &img,
+                                const nav_msgs::msg::Odometry::ConstPtr &odom)
 {
   /* get pose */
   Eigen::Quaterniond body_q = Eigen::Quaterniond(odom->pose.pose.orientation.w,
