@@ -80,7 +80,7 @@ TEST(GeometryUtilsFloat, Angle) {
 }
 
 TEST(ConverterDouble, Equality) {
-    nav_msgs::OdometryPtr pOdom(new nav_msgs::Odometry());
+    auto pOdom = std::make_shared<nav_msgs::msg::Odometry>();
 
     pOdom->pose.pose.position.x = 1.0;
     pOdom->pose.pose.position.y = 2.0;
@@ -102,7 +102,7 @@ TEST(ConverterDouble, Equality) {
     Eigen::Vector3d p, v, w;
     Eigen::Quaterniond q;
 
-    nav_msgs::Odometry odom_ = *pOdom;
+    nav_msgs::msg::Odometry odom_ = *pOdom;
 
     extract_odometry(pOdom, p, v, q, w);
 
