@@ -5,19 +5,32 @@
 namespace ego_planner
 {
 
-  void BsplineOptimizer::setParam(ros::NodeHandle &nh)
+  void BsplineOptimizer::setParam(rclcpp::Node::SharedPtr node)
   {
-    nh.param("optimization/lambda_smooth", lambda1_, -1.0);
-    nh.param("optimization/lambda_collision", lambda2_, -1.0);
-    nh.param("optimization/lambda_feasibility", lambda3_, -1.0);
-    nh.param("optimization/lambda_fitness", lambda4_, -1.0);
 
-    nh.param("optimization/dist0", dist0_, -1.0);
-    nh.param("optimization/swarm_clearance", swarm_clearance_, -1.0);
-    nh.param("optimization/max_vel", max_vel_, -1.0);
-    nh.param("optimization/max_acc", max_acc_, -1.0);
+    node->declare_parameter("optimization/lambda_smooth", -1.0);
+    node->declare_parameter("optimization/lambda_collision", -1.0);
+    node->declare_parameter("optimization/lambda_feasibility", -1.0);
+    node->declare_parameter("optimization/lambda_fitness", -1.0);
 
-    nh.param("optimization/order", order_, 3);
+    node->declare_parameter("optimization/dist0", -1.0);
+    node->declare_parameter("optimization/swarm_clearance", -1.0);
+    node->declare_parameter("optimization/max_vel", -1.0);
+    node->declare_parameter("optimization/max_acc", -1.0);
+
+    node->declare_parameter("optimization/order", 3);
+
+    node->get_parameter("optimization/lambda_smooth", lambda1_);
+    node->get_parameter("optimization/lambda_collision", lambda2_);
+    node->get_parameter("optimization/lambda_feasibility", lambda3_);
+    node->get_parameter("optimization/lambda_fitness", lambda4_);
+
+    node->get_parameter("optimization/dist0", dist0_);
+    node->get_parameter("optimization/swarm_clearance", swarm_clearance_);
+    node->get_parameter("optimization/max_vel", max_vel_);
+    node->get_parameter("optimization/max_acc", max_acc_);
+
+    node->get_parameter("optimization/order", order_);
   }
 
   void BsplineOptimizer::setEnvironment(const GridMap::Ptr &map)
@@ -42,6 +55,7 @@ namespace ego_planner
 
   void BsplineOptimizer::setDroneId(const int drone_id) { drone_id_ = drone_id; }
 
+  // 返回多个安全的控制点集
   std::vector<ControlPoints> BsplineOptimizer::distinctiveTrajs(vector<std::pair<int, int>> segments)
   {
     if (segments.size() == 0) // will be invoked again later.
@@ -51,20 +65,22 @@ namespace ego_planner
       return oneSeg;
     }
 
-    constexpr int MAX_TRAJS = 8;
-    constexpr int VARIS = 2;
-    int seg_upbound = std::min((int)segments.size(), static_cast<int>(floor(log(MAX_TRAJS) / log(VARIS))));
+    constexpr int MAX_TRAJS = 8;                                                                            // 最多的轨迹数量
+    constexpr int VARIS = 2;                                                                                // 允许的变化种类数
+    int seg_upbound = std::min((int)segments.size(), static_cast<int>(floor(log(MAX_TRAJS) / log(VARIS)))); // 允许变换的片段数量上限
     std::vector<ControlPoints> control_pts_buf;
     control_pts_buf.reserve(MAX_TRAJS);
     const double RESOLUTION = grid_map_->getResolution();
-    const double CTRL_PT_DIST = (cps_.points.col(0) - cps_.points.col(cps_.size - 1)).norm() / (cps_.size - 1);
+    const double CTRL_PT_DIST = (cps_.points.col(0) - cps_.points.col(cps_.size - 1)).norm() / (cps_.size - 1); // 计算控制点间的平均距离
 
     // Step 1. Find the opposite vectors and base points for every segment.
     std::vector<std::pair<ControlPoints, ControlPoints>> RichInfoSegs;
+    // 初始化两套控制点信息
     for (int i = 0; i < seg_upbound; i++)
     {
       std::pair<ControlPoints, ControlPoints> RichInfoOneSeg;
       ControlPoints RichInfoOneSeg_temp;
+      // 获取指定片段的控制点信息
       cps_.segment(RichInfoOneSeg_temp, segments[i].first, segments[i].second);
       RichInfoOneSeg.first = RichInfoOneSeg_temp;
       RichInfoOneSeg.second = RichInfoOneSeg_temp;
@@ -93,12 +109,14 @@ namespace ego_planner
         Eigen::Vector3d occ_start_pt, occ_end_pt;
         for (int j = 0; j < RichInfoSegs[i].first.size - 1; j++)
         {
-          //cout << "A *" << j << "*" << endl;
+          // cout << "A *" << j << "*" << endl;
+          //  遍历每个控制点及其后一个点，在两点间通过线性插值生成采样点
           double step_size = RESOLUTION / (RichInfoSegs[i].first.points.col(j) - RichInfoSegs[i].first.points.col(j + 1)).norm() / 2;
           for (double a = 1; a > 0; a -= step_size)
           {
             Eigen::Vector3d pt(a * RichInfoSegs[i].first.points.col(j) + (1 - a) * RichInfoSegs[i].first.points.col(j + 1));
-            //cout << " " << grid_map_->getInflateOccupancy(pt) << " pt=" << pt.transpose() << endl;
+            // cout << " " << grid_map_->getInflateOccupancy(pt) << " pt=" << pt.transpose() << endl;
+            //  如果检测到在障碍物内，则存储对应数据
             if (grid_map_->getInflateOccupancy(pt))
             {
               occ_start_id = j;
@@ -108,16 +126,18 @@ namespace ego_planner
           }
         }
       exit_multi_loop1:;
+        // 查找片段与最后一个障碍物的交点
         for (int j = RichInfoSegs[i].first.size - 1; j >= 1; j--)
         {
-          //cout << "j=" << j << endl;
-          //cout << "B *" << j << "*" << endl;
+          // cout << "j=" << j << endl;
+          // cout << "B *" << j << "*" << endl;
           ;
+          // 和上面同样的采样，然后检测是否在障碍物中
           double step_size = RESOLUTION / (RichInfoSegs[i].first.points.col(j) - RichInfoSegs[i].first.points.col(j - 1)).norm();
           for (double a = 1; a > 0; a -= step_size)
           {
             Eigen::Vector3d pt(a * RichInfoSegs[i].first.points.col(j) + (1 - a) * RichInfoSegs[i].first.points.col(j - 1));
-            //cout << " " << grid_map_->getInflateOccupancy(pt) << " pt=" << pt.transpose() << endl;
+            // cout << " " << grid_map_->getInflateOccupancy(pt) << " pt=" << pt.transpose() << endl;
             ;
             if (grid_map_->getInflateOccupancy(pt))
             {
@@ -130,6 +150,7 @@ namespace ego_planner
       exit_multi_loop2:;
 
         // double check
+        // 如果片段的起点或者终点在障碍物中，将会被移除
         if (occ_start_id == -1 || occ_end_id == -1)
         {
           // It means that the first or the last control points of one segment are in obstacles, which is not allowed.
@@ -160,19 +181,21 @@ namespace ego_planner
         for (int j = occ_start_id; j <= occ_end_id; j++)
         {
           Eigen::Vector3d base_pt_reverse, base_vec_reverse;
+          // 检查控制点的base point是否为1
           if (RichInfoSegs[i].first.base_point[j].size() != 1)
           {
             cout << "RichInfoSegs[" << i << "].first.base_point[" << j << "].size()=" << RichInfoSegs[i].first.base_point[j].size() << endl;
-            ROS_ERROR("Wrong number of base_points!!! Should not be happen!.");
+            RCLCPP_ERROR(rclcpp::get_logger("distinctiveTrajs"), "Wrong number of base_points!!! Should not be happen!.");
 
             cout << setprecision(5);
             cout << "cps_" << endl;
             cout << " clearance=" << cps_.clearance << " cps.size=" << cps_.size << endl;
+            // 输出错误信息
             for (int temp_i = 0; temp_i < cps_.size; temp_i++)
             {
               if (cps_.base_point[temp_i].size() > 1 && cps_.base_point[temp_i].size() < 1000)
               {
-                ROS_ERROR("Should not happen!!!");
+                RCLCPP_ERROR(rclcpp::get_logger("distinctiveTrajs"), "Should not happen!!!");
                 cout << "######" << cps_.points.col(temp_i).transpose() << endl;
                 for (size_t temp_j = 0; temp_j < cps_.base_point[temp_i].size(); temp_j++)
                   cout << "      " << cps_.base_point[temp_i][temp_j].transpose() << " @ " << cps_.direction[temp_i][temp_j].transpose() << endl;
@@ -183,30 +206,37 @@ namespace ego_planner
             return blank;
           }
 
+          // 通过取反获得相反方向的向量
           base_vec_reverse = -RichInfoSegs[i].first.direction[j][0];
 
           // The start and the end case must get taken special care of.
+          // 若当前控制点为片段的起始点 occ_start_id，则将障碍物交点 occ_start_pt 直接设为 base_pt_reverse
           if (j == occ_start_id)
           {
             base_pt_reverse = occ_start_pt;
           }
+          // 若当前控制点为片段的终止点 occ_end_id，则将终点交点 occ_end_pt 设为 base_pt_reverse
           else if (j == occ_end_id)
           {
             base_pt_reverse = occ_end_pt;
           }
+          // 对于片段中的中间控制点，将基准点 base_pt_reverse 设置为当前控制点 points.col(j) 沿反向向量 base_vec_reverse 方向延伸的某一距离位置
           else
           {
             base_pt_reverse = RichInfoSegs[i].first.points.col(j) + base_vec_reverse * (RichInfoSegs[i].first.base_point[j][0] - RichInfoSegs[i].first.points.col(j)).norm();
           }
 
+          // 检查base_pt_reverse是否在障碍物中
           if (grid_map_->getInflateOccupancy(base_pt_reverse)) // Search outward.
           {
+            // 最大搜索范围
             double l_upbound = 5 * CTRL_PT_DIST; // "5" is the threshold.
             double l = RESOLUTION;
             for (; l <= l_upbound; l += RESOLUTION)
             {
+              // 不断将控制点向外移动，寻找不在障碍物中的控制点
               Eigen::Vector3d base_pt_temp = base_pt_reverse + l * base_vec_reverse;
-              //cout << base_pt_temp.transpose() << endl;
+              // cout << base_pt_temp.transpose() << endl;
               if (!grid_map_->getInflateOccupancy(base_pt_temp))
               {
                 RichInfoSegs[i].second.base_point[j][0] = base_pt_temp;
@@ -214,9 +244,10 @@ namespace ego_planner
                 break;
               }
             }
+            // 如果找不到则删除这一段
             if (l > l_upbound)
             {
-              ROS_WARN("Can't find the new base points at the opposite within the threshold. i=%d, j=%d", i, j);
+              RCLCPP_WARN(rclcpp::get_logger("distinctiveTrajs"), "Can't find the new base points at the opposite within the threshold. i=%d, j=%d", i, j);
 
               segments.erase(segments.begin() + i);
               RichInfoSegs.erase(RichInfoSegs.begin() + i);
@@ -226,14 +257,16 @@ namespace ego_planner
               goto exit_multi_loop3; // break "for (int j = 0; j < RichInfoSegs[i].first.size; j++)"
             }
           }
+          // 如果距离控制点足够远且不再障碍物中则无需继续搜索
           else if ((base_pt_reverse - RichInfoSegs[i].first.points.col(j)).norm() >= RESOLUTION) // Unnecessary to search.
           {
             RichInfoSegs[i].second.base_point[j][0] = base_pt_reverse;
             RichInfoSegs[i].second.direction[j][0] = base_vec_reverse;
           }
+          // 基点和控制点太近则删除这一段
           else
           {
-            ROS_WARN("base_point and control point are too close!");
+            RCLCPP_WARN(rclcpp::get_logger("distinctiveTrajs"), "base_point and control point are too close!");
             cout << "base_point=" << RichInfoSegs[i].first.base_point[j][0].transpose() << " control point=" << RichInfoSegs[i].first.points.col(j).transpose() << endl;
 
             segments.erase(segments.begin() + i);
@@ -248,6 +281,7 @@ namespace ego_planner
         // 1.3 Assign the base points to control points within [0, occ_start_id) and (occ_end_id, RichInfoSegs[i].first.size()-1].
         if (RichInfoSegs[i].second.size)
         {
+          // 为片段起点之前和终点之后的控制点设置统一的基准点和方向，使得这些控制点在障碍物影响范围外时能够保持一致的路径属性
           for (int j = occ_start_id - 1; j >= 0; j--)
           {
             RichInfoSegs[i].second.base_point[j][0] = RichInfoSegs[i].second.base_point[occ_start_id][0];
@@ -262,6 +296,7 @@ namespace ego_planner
 
       exit_multi_loop3:;
       }
+      // 片段只有一个控制点的情况
       else if (RichInfoSegs[i].first.size == 1)
       {
         cout << "i=" << i << " RichInfoSegs.size()=" << RichInfoSegs.size() << endl;
@@ -281,7 +316,7 @@ namespace ego_planner
           for (; l <= l_upbound; l += RESOLUTION)
           {
             Eigen::Vector3d base_pt_temp = base_pt_reverse + l * base_vec_reverse;
-            //cout << base_pt_temp.transpose() << endl;
+            // cout << base_pt_temp.transpose() << endl;
             if (!grid_map_->getInflateOccupancy(base_pt_temp))
             {
               RichInfoSegs[i].second.base_point[0][0] = base_pt_temp;
@@ -291,7 +326,8 @@ namespace ego_planner
           }
           if (l > l_upbound)
           {
-            ROS_WARN("Can't find the new base points at the opposite within the threshold, 2. i=%d", i);
+            RCLCPP_WARN(rclcpp::get_logger("distinctiveTrajs"), 
+                        "Can't find the new base points at the opposite within the threshold, 2. i=%d", i);
 
             segments.erase(segments.begin() + i);
             RichInfoSegs.erase(RichInfoSegs.begin() + i);
@@ -306,7 +342,8 @@ namespace ego_planner
         }
         else
         {
-          ROS_WARN("base_point and control point are too close!, 2");
+          RCLCPP_WARN(rclcpp::get_logger("distinctiveTrajs"), 
+                        "base_point and control point are too close!, 2");
           cout << "base_point=" << RichInfoSegs[i].first.base_point[0][0].transpose() << " control point=" << RichInfoSegs[i].first.points.col(0).transpose() << endl;
 
           segments.erase(segments.begin() + i);
@@ -326,6 +363,7 @@ namespace ego_planner
     // cout << "A3" << endl;
 
     // Step 2. Assemble each segment to make up the new control point sequence.
+    // 将每个分段组合起来，组成新的控制点序列
     if (seg_upbound == 0) // After the erase operation above, segment legth will decrease to 0 again.
     {
       std::vector<ControlPoints> oneSeg;
@@ -333,22 +371,27 @@ namespace ego_planner
       return oneSeg;
     }
 
+    // 初始化选择向量
     std::vector<int> selection(seg_upbound);
     std::fill(selection.begin(), selection.end(), 0);
     selection[0] = -1; // init
+    // 计算最大组合数
     int max_traj_nums = static_cast<int>(pow(VARIS, seg_upbound));
     for (int i = 0; i < max_traj_nums; i++)
     {
       // 2.1 Calculate the selection table.
       int digit_id = 0;
       selection[digit_id]++;
+      // 生成一个选择表
       while (digit_id < seg_upbound && selection[digit_id] >= VARIS)
       {
         selection[digit_id] = 0;
         digit_id++;
         if (digit_id >= seg_upbound)
         {
-          ROS_ERROR("Should not happen!!! digit_id=%d, seg_upbound=%d", digit_id, seg_upbound);
+          RCLCPP_ERROR(rclcpp::get_logger("distinctiveTrajs"), 
+                        "Should not happen!!! digit_id=%d, seg_upbound=%d", digit_id, seg_upbound);
+          
         }
         selection[digit_id]++;
       }
@@ -358,28 +401,31 @@ namespace ego_planner
       cpsOneSample.resize(cps_.size);
       cpsOneSample.clearance = cps_.clearance;
       int cp_id = 0, seg_id = 0, cp_of_seg_id = 0;
+      // 遍历所有控制点
       while (/*seg_id < RichInfoSegs.size() ||*/ cp_id < cps_.size)
       {
-        //cout << "A ";
-        // if ( seg_id >= RichInfoSegs.size() )
-        // {
-        //   cout << "seg_id=" << seg_id << " RichInfoSegs.size()=" << RichInfoSegs.size() << endl;
-        // }
-        // if ( cp_id >= cps_.base_point.size() )
-        // {
-        //   cout << "cp_id=" << cp_id << " cps_.base_point.size()=" << cps_.base_point.size() << endl;
-        // }
-        // if ( cp_of_seg_id >= RichInfoSegs[seg_id].first.base_point.size() )
-        // {
-        //   cout << "cp_of_seg_id=" << cp_of_seg_id << " RichInfoSegs[seg_id].first.base_point.size()=" << RichInfoSegs[seg_id].first.base_point.size() << endl;
-        // }
-
+        // cout << "A ";
+        //  if ( seg_id >= RichInfoSegs.size() )
+        //  {
+        //    cout << "seg_id=" << seg_id << " RichInfoSegs.size()=" << RichInfoSegs.size() << endl;
+        //  }
+        //  if ( cp_id >= cps_.base_point.size() )
+        //  {
+        //    cout << "cp_id=" << cp_id << " cps_.base_point.size()=" << cps_.base_point.size() << endl;
+        //  }
+        //  if ( cp_of_seg_id >= RichInfoSegs[seg_id].first.base_point.size() )
+        //  {
+        //    cout << "cp_of_seg_id=" << cp_of_seg_id << " RichInfoSegs[seg_id].first.base_point.size()=" << RichInfoSegs[seg_id].first.base_point.size() << endl;
+        //  }
+        //  判断控制点是否在当前控制范围内
+        //  如果不在则直接从原始控制点集中复制数据
         if (seg_id >= seg_upbound || cp_id < segments[seg_id].first || cp_id > segments[seg_id].second)
         {
           cpsOneSample.points.col(cp_id) = cps_.points.col(cp_id);
           cpsOneSample.base_point[cp_id] = cps_.base_point[cp_id];
           cpsOneSample.direction[cp_id] = cps_.direction[cp_id];
         }
+        // 如果 cp_id 位于当前片段范围内，根据 selection[seg_id] 的值选择片段的第一套或第二套基准点和方向
         else if (cp_id >= segments[seg_id].first && cp_id <= segments[seg_id].second)
         {
           if (!selection[seg_id]) // zx-todo
@@ -405,6 +451,7 @@ namespace ego_planner
             }
           }
 
+          // 当遍历到片段的最后一个控制点时，将 cp_of_seg_id 重置为 0，并将 seg_id 指向下一个片段
           if (cp_id == segments[seg_id].second)
           {
             cp_of_seg_id = 0;
@@ -413,7 +460,8 @@ namespace ego_planner
         }
         else
         {
-          ROS_ERROR("Shold not happen!!!!, cp_id=%d, seg_id=%d, segments.front().first=%d, segments.back().second=%d, segments[seg_id].first=%d, segments[seg_id].second=%d",
+          RCLCPP_ERROR(rclcpp::get_logger("distinctiveTrajs"), 
+                    "Shold not happen!!!!, cp_id=%d, seg_id=%d, segments.front().first=%d, segments.back().second=%d, segments[seg_id].first=%d, segments[seg_id].second=%d",
                     cp_id, seg_id, segments.front().first, segments.back().second, segments[seg_id].first, segments[seg_id].second);
         }
 
@@ -428,9 +476,10 @@ namespace ego_planner
     return control_pts_buf;
   } // namespace ego_planner
 
-  /* This function is very similar to check_collision_and_rebound(). 
+  /* This function is very similar to check_collision_and_rebound().
    * It was written separately, just because I did it once and it has been running stably since March 2020.
    * But I will merge then someday.*/
+  // 初始化控制点
   std::vector<std::pair<int, int>> BsplineOptimizer::initControlPoints(Eigen::MatrixXd &init_points, bool flag_first_init /*= true*/)
   {
 
@@ -442,24 +491,31 @@ namespace ego_planner
     }
 
     /*** Segment the initial trajectory according to obstacles ***/
+    // 进入或离开障碍物稳定的时间间隔
     constexpr int ENOUGH_INTERVAL = 2;
+    // 障碍物检测的步长
     double step_size = grid_map_->getResolution() / ((init_points.col(0) - init_points.rightCols(1)).norm() / (init_points.cols() - 1)) / 1.5;
     int in_id = -1, out_id = -1;
     vector<std::pair<int, int>> segment_ids;
     int same_occ_state_times = ENOUGH_INTERVAL + 1;
     bool occ, last_occ = false;
+    // 标识片段的起点和终点是否找到
     bool flag_got_start = false, flag_got_end = false, flag_got_end_maybe = false;
     int i_end = (int)init_points.cols() - order_ - ((int)init_points.cols() - 2 * order_) / 3; // only check closed 2/3 points.
+    // 遍历所有点
     for (int i = order_; i <= i_end; ++i)
     {
-      //cout << " *" << i-1 << "*" ;
+      // cout << " *" << i-1 << "*" ;
+      //  相邻两个点之间进行线性插值并检测障碍物
       for (double a = 1.0; a > 0.0; a -= step_size)
       {
+        // TODO:没搞懂这是干嘛的
         occ = grid_map_->getInflateOccupancy(a * init_points.col(i - 1) + (1 - a) * init_points.col(i));
-        //cout << " " << occ;
-        // cout << setprecision(5);
-        // cout << (a * init_points.col(i-1) + (1-a) * init_points.col(i)).transpose() << " occ1=" << occ << endl;
+        // cout << " " << occ;
+        //  cout << setprecision(5);
+        //  cout << (a * init_points.col(i-1) + (1-a) * init_points.col(i)).transpose() << " occ1=" << occ << endl;
 
+        // 进入障碍物
         if (occ && !last_occ)
         {
           if (same_occ_state_times > ENOUGH_INTERVAL || i == order_)
@@ -470,17 +526,20 @@ namespace ego_planner
           same_occ_state_times = 0;
           flag_got_end_maybe = false; // terminate in advance
         }
+        // 离开障碍物
         else if (!occ && last_occ)
         {
           out_id = i;
           flag_got_end_maybe = true;
           same_occ_state_times = 0;
         }
+        // 如果状态没发生变化
         else
         {
           ++same_occ_state_times;
         }
 
+        // 如果已经离开障碍物，则结束
         if (flag_got_end_maybe && (same_occ_state_times > ENOUGH_INTERVAL || (i == (int)init_points.cols() - order_)))
         {
           flag_got_end_maybe = false;
@@ -489,6 +548,7 @@ namespace ego_planner
 
         last_occ = occ;
 
+        // 重置标志位并存储信息
         if (flag_got_start && flag_got_end)
         {
           flag_got_start = false;
@@ -512,10 +572,11 @@ namespace ego_planner
     }
 
     /*** a star search ***/
+    // 在每个无障碍片段 segment_ids 的起点和终点之间寻找一条路径
     vector<vector<Eigen::Vector3d>> a_star_pathes;
     for (size_t i = 0; i < segment_ids.size(); ++i)
     {
-      //cout << "in=" << in.transpose() << " out=" << out.transpose() << endl;
+      // cout << "in=" << in.transpose() << " out=" << out.transpose() << endl;
       Eigen::Vector3d in(init_points.col(segment_ids[i].first)), out(init_points.col(segment_ids[i].second));
       if (a_star_->AstarSearch(/*(in-out).norm()/10+0.05*/ 0.1, in, out))
       {
@@ -523,7 +584,7 @@ namespace ego_planner
       }
       else
       {
-        ROS_ERROR("a star error, force return!");
+        RCLCPP_ERROR(rclcpp::get_logger("initControlPoints"), "a star error, force return!");
         vector<std::pair<int, int>> blank_ret;
         return blank_ret;
       }
@@ -532,6 +593,7 @@ namespace ego_planner
     /*** calculate bounds ***/
     int id_low_bound, id_up_bound;
     vector<std::pair<int, int>> bounds(segment_ids.size());
+    // 遍历每一段，设置边界
     for (size_t i = 0; i < segment_ids.size(); i++)
     {
 
@@ -540,22 +602,29 @@ namespace ego_planner
         id_low_bound = order_;
         if (segment_ids.size() > 1)
         {
+          // 取当前片段结束点 segment_ids[0].second 和下一个片段起始点 segment_ids[1].first 的中间位置，向下取整后作为高边界
           id_up_bound = (int)(((segment_ids[0].second + segment_ids[1].first) - 1.0f) / 2); // id_up_bound : -1.0f fix()
         }
         else
         {
+          // 距末尾 order_ + 1 个点的位置
           id_up_bound = init_points.cols() - order_ - 1;
         }
       }
+      // 末尾的边界
       else if (i == segment_ids.size() - 1) // last segment, i != 0 here
       {
+        // 尾段的低边界为当前片段起点和上一个片段终点的中间位置，向上取整
         id_low_bound = (int)(((segment_ids[i].first + segment_ids[i - 1].second) + 1.0f) / 2); // id_low_bound : +1.0f ceil()
+        // 距末尾点 order_ + 1 个点
         id_up_bound = init_points.cols() - order_ - 1;
       }
       else
       {
+        // 低边界：为当前片段起点和前一片段终点的中间位置，向上取整
         id_low_bound = (int)(((segment_ids[i].first + segment_ids[i - 1].second) + 1.0f) / 2); // id_low_bound : +1.0f ceil()
-        id_up_bound = (int)(((segment_ids[i].second + segment_ids[i + 1].first) - 1.0f) / 2);  // id_up_bound : -1.0f fix()
+        // 高边界：为当前片段终点和下一片段起点的中间位置，向下取整
+        id_up_bound = (int)(((segment_ids[i].second + segment_ids[i + 1].first) - 1.0f) / 2); // id_up_bound : -1.0f fix()
       }
 
       bounds[i] = std::pair<int, int>(id_low_bound, id_up_bound);
@@ -569,15 +638,19 @@ namespace ego_planner
 
     /*** Adjust segment length ***/
     vector<std::pair<int, int>> adjusted_segment_ids(segment_ids.size());
+    // 控制点的最小比例
     constexpr double MINIMUM_PERCENT = 0.0; // Each segment is guaranteed to have sufficient points to generate sufficient force
+    // 控制点的最小数量
     int minimum_points = round(init_points.cols() * MINIMUM_PERCENT), num_points;
     for (size_t i = 0; i < segment_ids.size(); i++)
     {
       /*** Adjust segment length ***/
+      // 获取当前片段的点数
       num_points = segment_ids[i].second - segment_ids[i].first + 1;
-      //cout << "i = " << i << " first = " << segment_ids[i].first << " second = " << segment_ids[i].second << endl;
+      // cout << "i = " << i << " first = " << segment_ids[i].first << " second = " << segment_ids[i].second << endl;
       if (num_points < minimum_points)
       {
+        // 如果点数不够则在两侧扩展点，确保不能超过边界
         double add_points_each_side = (int)(((minimum_points - num_points) + 1.0f) / 2);
 
         adjusted_segment_ids[i].first = segment_ids[i].first - add_points_each_side >= bounds[i].first ? segment_ids[i].first - add_points_each_side : bounds[i].first;
@@ -590,8 +663,9 @@ namespace ego_planner
         adjusted_segment_ids[i].second = segment_ids[i].second;
       }
 
-      //cout << "final:" << "i = " << i << " first = " << adjusted_segment_ids[i].first << " second = " << adjusted_segment_ids[i].second << endl;
+      // cout << "final:" << "i = " << i << " first = " << adjusted_segment_ids[i].first << " second = " << adjusted_segment_ids[i].second << endl;
     }
+    // 避免重叠
     for (size_t i = 1; i < adjusted_segment_ids.size(); i++) // Avoid overlap
     {
       if (adjusted_segment_ids[i - 1].second >= adjusted_segment_ids[i].first)
@@ -609,20 +683,26 @@ namespace ego_planner
     for (size_t i = 0; i < segment_ids.size(); i++)
     {
       // step 1
+      // 遍历该段所有控制点的id，并将标志位置为false
       for (int j = adjusted_segment_ids[i].first; j <= adjusted_segment_ids[i].second; ++j)
         cps_.flag_temp[j] = false;
 
       // step 2
+      // 初始化交点标记
       int got_intersection_id = -1;
       for (int j = segment_ids[i].first + 1; j < segment_ids[i].second; ++j)
       {
+        // 计算控制点的方向向量
         Eigen::Vector3d ctrl_pts_law(init_points.col(j + 1) - init_points.col(j - 1)), intersection_point;
+        // A*路径的中点
         int Astar_id = a_star_pathes[i].size() / 2, last_Astar_id; // Let "Astar_id = id_of_the_most_far_away_Astar_point" will be better, but it needs more computation
+        // 路径点和控制点之间的点积，判断当前路径点在控制点方向 ctrl_pts_law 的哪一侧
         double val = (a_star_pathes[i][Astar_id] - init_points.col(j)).dot(ctrl_pts_law), last_val = val;
         while (Astar_id >= 0 && Astar_id < (int)a_star_pathes[i].size())
         {
           last_Astar_id = Astar_id;
 
+          // 根据 val 的正负决定 Astar_id 的移动方向
           if (val >= 0)
             --Astar_id;
           else
@@ -632,13 +712,14 @@ namespace ego_planner
 
           if (val * last_val <= 0 && (abs(val) > 0 || abs(last_val) > 0)) // val = last_val = 0.0 is not allowed
           {
+            // 寻找可能的交点，TODO:还没搞懂
             intersection_point =
                 a_star_pathes[i][Astar_id] +
                 ((a_star_pathes[i][Astar_id] - a_star_pathes[i][last_Astar_id]) *
                  (ctrl_pts_law.dot(init_points.col(j) - a_star_pathes[i][Astar_id]) / ctrl_pts_law.dot(a_star_pathes[i][Astar_id] - a_star_pathes[i][last_Astar_id])) // = t
                 );
 
-            //cout << "i=" << i << " j=" << j << " Astar_id=" << Astar_id << " last_Astar_id=" << last_Astar_id << " intersection_point = " << intersection_point.transpose() << endl;
+            // cout << "i=" << i << " j=" << j << " Astar_id=" << Astar_id << " last_Astar_id=" << last_Astar_id << " intersection_point = " << intersection_point.transpose() << endl;
 
             got_intersection_id = j;
             break;
@@ -647,18 +728,22 @@ namespace ego_planner
 
         if (got_intersection_id >= 0)
         {
+          // 计算交点到控制点之间的距离
           double length = (intersection_point - init_points.col(j)).norm();
           if (length > 1e-5)
           {
             cps_.flag_temp[j] = true;
+            // 逐步进行采样
             for (double a = length; a >= 0.0; a -= grid_map_->getResolution())
             {
+              // 通过线性插值计算采样点位置
               occ = grid_map_->getInflateOccupancy((a / length) * intersection_point + (1 - a / length) * init_points.col(j));
 
               if (occ || a < grid_map_->getResolution())
               {
                 if (occ)
                   a += grid_map_->getResolution();
+                // 记录基点并计算到控制点的方向
                 cps_.base_point[j].push_back((a / length) * intersection_point + (1 - a / length) * init_points.col(j));
                 cps_.direction[j].push_back((intersection_point - init_points.col(j)).normalized());
                 // cout << "A " << j << endl;
@@ -674,10 +759,13 @@ namespace ego_planner
       }
 
       /* Corner case: the segment length is too short. Here the control points may outside the A* path, leading to opposite gradient direction. So I have to take special care of it */
+      // 当轨迹片段只有两个控制点（即长度过短）时，A* 路径的方向可能与控制点的方向不一致，导致梯度方向相反
       if (segment_ids[i].second - segment_ids[i].first == 1)
       {
+        // 计算控制点方向向量和中点
         Eigen::Vector3d ctrl_pts_law(init_points.col(segment_ids[i].second) - init_points.col(segment_ids[i].first)), intersection_point;
         Eigen::Vector3d middle_point = (init_points.col(segment_ids[i].second) + init_points.col(segment_ids[i].first)) / 2;
+        // 计算A*中点并判断方向，同上
         int Astar_id = a_star_pathes[i].size() / 2, last_Astar_id; // Let "Astar_id = id_of_the_most_far_away_Astar_point" will be better, but it needs more computation
         double val = (a_star_pathes[i][Astar_id] - middle_point).dot(ctrl_pts_law), last_val = val;
         while (Astar_id >= 0 && Astar_id < (int)a_star_pathes[i].size())
@@ -689,16 +777,19 @@ namespace ego_planner
           else
             ++Astar_id;
 
+          // 和上面不一样，这里减去的是中点
           val = (a_star_pathes[i][Astar_id] - middle_point).dot(ctrl_pts_law);
 
           if (val * last_val <= 0 && (abs(val) > 0 || abs(last_val) > 0)) // val = last_val = 0.0 is not allowed
           {
+            // 计算交点
             intersection_point =
                 a_star_pathes[i][Astar_id] +
                 ((a_star_pathes[i][Astar_id] - a_star_pathes[i][last_Astar_id]) *
                  (ctrl_pts_law.dot(middle_point - a_star_pathes[i][Astar_id]) / ctrl_pts_law.dot(a_star_pathes[i][Astar_id] - a_star_pathes[i][last_Astar_id])) // = t
                 );
 
+            // 满足距离要求则存储相关信息
             if ((intersection_point - middle_point).norm() > 0.01) // 1cm.
             {
               cps_.flag_temp[segment_ids[i].first] = true;
@@ -712,10 +803,12 @@ namespace ego_planner
         }
       }
 
-      //step 3
+      // step 3
       if (got_intersection_id >= 0)
       {
+        // 遍历交点之后的控制点
         for (int j = got_intersection_id + 1; j <= adjusted_segment_ids[i].second; ++j)
+          // 如果没被标记则将前一个控制点的信息给该点
           if (!cps_.flag_temp[j])
           {
             cps_.base_point[j].push_back(cps_.base_point[j - 1].back());
@@ -723,6 +816,7 @@ namespace ego_planner
             // cout << "AAA " << j << endl;
           }
 
+        // 遍历交点前的控制点，如果没被标记用后一个控制点的数据赋值
         for (int j = got_intersection_id - 1; j >= adjusted_segment_ids[i].first; --j)
           if (!cps_.flag_temp[j])
           {
@@ -743,6 +837,7 @@ namespace ego_planner
     return final_segment_ids;
   }
 
+  // 急停情况下提前退出
   int BsplineOptimizer::earlyExit(void *func_data, const double *x, const double *g, const double fx, const double xnorm, const double gnorm, const double step, int n, int k, int ls)
   {
     BsplineOptimizer *opt = reinterpret_cast<BsplineOptimizer *>(func_data);
@@ -751,6 +846,7 @@ namespace ego_planner
     return (opt->force_stop_type_ == STOP_FOR_ERROR || opt->force_stop_type_ == STOP_FOR_REBOUND);
   }
 
+  // 利用combineCostRebound计算损失
   double BsplineOptimizer::costFunctionRebound(void *func_data, const double *x, double *grad, const int n)
   {
     BsplineOptimizer *opt = reinterpret_cast<BsplineOptimizer *>(func_data);
@@ -762,6 +858,7 @@ namespace ego_planner
     return cost;
   }
 
+  // 利用combineCostRefine计算优化后的损失
   double BsplineOptimizer::costFunctionRefine(void *func_data, const double *x, double *grad, const int n)
   {
     BsplineOptimizer *opt = reinterpret_cast<BsplineOptimizer *>(func_data);
@@ -773,12 +870,13 @@ namespace ego_planner
     return cost;
   }
 
+  // 几个计算损失的函数
   void BsplineOptimizer::calcSwarmCost(const Eigen::MatrixXd &q, double &cost, Eigen::MatrixXd &gradient)
   {
     cost = 0.0;
     int end_idx = q.cols() - order_ - (double)(q.cols() - 2 * order_) * 1.0 / 3.0; // Only check the first 2/3 points
     const double CLEARANCE = swarm_clearance_ * 2;
-    double t_now = ros::Time::now().toSec();
+    double t_now = rclcpp::Clock().now().seconds();
     constexpr double a = 2.0, b = 1.0, inv_a2 = 1 / a / a, inv_b2 = 1 / b / b;
 
     for (int i = order_; i < end_idx; i++)
@@ -792,7 +890,7 @@ namespace ego_planner
           continue;
         }
 
-        double traj_i_satrt_time = swarm_trajs_->at(id).start_time_.toSec();
+        double traj_i_satrt_time = swarm_trajs_->at(id).start_time_.seconds();
         if (glb_time < traj_i_satrt_time + swarm_trajs_->at(id).duration_ - 0.1)
         {
           /* def cost=(c-sqrt([Q-O]'D[Q-O]))^2, D=[1/b^2,0,0;0,1/b^2,0;0,0,1/a^2] */
@@ -831,7 +929,7 @@ namespace ego_planner
     cost = 0.0;
     int end_idx = q.cols() - order_;
     constexpr double CLEARANCE = 1.5;
-    double t_now = ros::Time::now().toSec();
+    double t_now = rclcpp::Clock().now().seconds();
 
     for (int i = order_; i < end_idx; i++)
     {
@@ -841,7 +939,7 @@ namespace ego_planner
       {
         Eigen::Vector3d obj_prid = moving_objs_->evaluateConstVel(id, t_now + time);
         double dist = (cps_.points.col(i) - obj_prid).norm();
-        //cout /*<< "cps_.points.col(i)=" << cps_.points.col(i).transpose()*/ << " moving_objs_=" << obj_prid.transpose() << " dist=" << dist << endl;
+        // cout /*<< "cps_.points.col(i)=" << cps_.points.col(i).transpose()*/ << " moving_objs_=" << obj_prid.transpose() << " dist=" << dist << endl;
         double dist_err = CLEARANCE - dist;
         Eigen::Vector3d dist_grad = (cps_.points.col(i) - obj_prid).normalized();
 
@@ -995,7 +1093,7 @@ namespace ego_planner
                                              Eigen::MatrixXd &gradient)
   {
 
-    //#define SECOND_DERIVATIVE_CONTINOUS
+    // #define SECOND_DERIVATIVE_CONTINOUS
 
 #ifdef SECOND_DERIVATIVE_CONTINOUS
 
@@ -1132,7 +1230,7 @@ namespace ego_planner
     {
       Eigen::Vector3d vi = (q.col(i + 1) - q.col(i)) / ts;
 
-      //cout << "temp_v * vi=" ;
+      // cout << "temp_v * vi=" ;
       for (int j = 0; j < 3; j++)
       {
         if (vi(j) > max_vel_)
@@ -1163,7 +1261,7 @@ namespace ego_planner
     {
       Eigen::Vector3d ai = (q.col(i + 2) - 2 * q.col(i + 1) + q.col(i)) * ts_inv2;
 
-      //cout << "temp_a * ai=" ;
+      // cout << "temp_a * ai=" ;
       for (int j = 0; j < 3; j++)
       {
         if (ai(j) > max_acc_)
@@ -1189,12 +1287,14 @@ namespace ego_planner
           /* code */
         }
       }
-      //cout << endl;
+      // cout << endl;
     }
 
 #endif
   }
+  /*                    上面七个都是计算损失的函数           */
 
+  // 检查是否有障碍物？
   bool BsplineOptimizer::check_collision_and_rebound(void)
   {
 
@@ -1240,7 +1340,7 @@ namespace ego_planner
         }
         if (j < 0) // fail to get the obs free point
         {
-          ROS_ERROR("ERROR! the drone is in obstacle. This should not happen.");
+          RCLCPP_ERROR(rclcpp::get_logger("check_collision_and_rebound"), "ERROR! the drone is in obstacle. This should not happen.");
           in_id = 0;
         }
 
@@ -1256,7 +1356,8 @@ namespace ego_planner
         }
         if (j >= cps_.size) // fail to get the obs free point
         {
-          ROS_WARN("WARN! terminal point of the current trajectory is in obstacle, skip this planning.");
+          RCLCPP_WARN(rclcpp::get_logger("check_collision_and_rebound"), 
+                      "WARN! terminal point of the current trajectory is in obstacle, skip this planning.");
 
           force_stop_type_ = STOP_FOR_ERROR;
           return false;
@@ -1281,7 +1382,7 @@ namespace ego_planner
         }
         else
         {
-          ROS_ERROR("a star error");
+          RCLCPP_ERROR(rclcpp::get_logger("check_collision_and_rebound"), "a star error");
           segment_ids.erase(segment_ids.begin() + i);
           i--;
         }
@@ -1364,7 +1465,7 @@ namespace ego_planner
           }
         }
 
-        //step 3
+        // step 3
         if (got_intersection_id >= 0)
         {
           for (int j = got_intersection_id + 1; j <= segment_ids[i].second; ++j)
@@ -1382,7 +1483,7 @@ namespace ego_planner
             }
         }
         else
-          ROS_WARN("Failed to generate direction. It doesn't matter.");
+          RCLCPP_WARN(rclcpp::get_logger("check_collision_and_rebound"), "Failed to generate direction. It doesn't matter.");
       }
 
       force_stop_type_ = STOP_FOR_REBOUND;
@@ -1392,6 +1493,7 @@ namespace ego_planner
     return false;
   }
 
+  // 设置时间间隔ts，调用rebound_optimize(final_cost)将轨迹推出障碍物，得到最优的无碰撞轨迹，并将其控制点赋值给optimal_points
   bool BsplineOptimizer::BsplineOptimizeTrajRebound(Eigen::MatrixXd &optimal_points, double ts)
   {
     setBsplineInterval(ts);
@@ -1404,8 +1506,11 @@ namespace ego_planner
     return flag_success;
   }
 
+  // 设置初始控制点control_points、时间间隔ts，调用rebound_optimize(final_cost)将轨迹推出障碍物，
+  // 得到最优的无碰撞轨迹，并将其控制点赋值给optimal_points
   bool BsplineOptimizer::BsplineOptimizeTrajRebound(Eigen::MatrixXd &optimal_points, double &final_cost, const ControlPoints &control_points, double ts)
   {
+    // 将时间间隔存储到成员变量
     setBsplineInterval(ts);
 
     cps_ = control_points;
@@ -1417,9 +1522,11 @@ namespace ego_planner
     return flag_success;
   }
 
+  // 设置初始控制点init_points、时间间隔ts，调用refine_optimize()重新分配时间，得到最优的动力学可行轨迹，并将其控制点赋值给optimal_points
   bool BsplineOptimizer::BsplineOptimizeTrajRefine(const Eigen::MatrixXd &init_points, const double ts, Eigen::MatrixXd &optimal_points)
   {
 
+    // 将控制点数据存储到成员变量
     setControlPoints(init_points);
     setBsplineInterval(ts);
 
@@ -1430,15 +1537,17 @@ namespace ego_planner
     return flag_success;
   }
 
+  // 使用L-BFGS方法对目标函数进行优化，得到光滑、无碰撞、动力学可行、与其他无人机碰撞、结束项的轨迹。
   bool BsplineOptimizer::rebound_optimize(double &final_cost)
   {
     iter_num_ = 0;
     int start_id = order_;
     // int end_id = this->cps_.size - order_; //Fixed end
     int end_id = this->cps_.size; // Free end
+    // 变量个数
     variable_num_ = 3 * (end_id - start_id);
 
-    ros::Time t0 = ros::Time::now(), t1, t2;
+    rclcpp::Time t0 = rclcpp::Clock().now(), t1, t2;
     int restart_nums = 0, rebound_times = 0;
     ;
     bool flag_force_return, flag_occ, success;
@@ -1454,29 +1563,33 @@ namespace ego_planner
       flag_occ = false;
       success = false;
 
+      // 控制点数组初始化
       double q[variable_num_];
       memcpy(q, cps_.points.data() + 3 * start_id, variable_num_ * sizeof(q[0]));
 
+      // 初始化L-BFGS算法的参数
       lbfgs::lbfgs_parameter_t lbfgs_params;
       lbfgs::lbfgs_load_default_parameters(&lbfgs_params);
-      lbfgs_params.mem_size = 16;
+      lbfgs_params.mem_size = 16; // 算法保存的历史优化步数
       lbfgs_params.max_iterations = 200;
       lbfgs_params.g_epsilon = 0.01;
 
       /* ---------- optimize ---------- */
-      t1 = ros::Time::now();
+      t1 = rclcpp::Clock().now();
+      // 执行优化
       int result = lbfgs::lbfgs_optimize(variable_num_, q, &final_cost, BsplineOptimizer::costFunctionRebound, NULL, BsplineOptimizer::earlyExit, this, &lbfgs_params);
-      t2 = ros::Time::now();
-      double time_ms = (t2 - t1).toSec() * 1000;
-      double total_time_ms = (t2 - t0).toSec() * 1000;
+      t2 = rclcpp::Clock().now();
+      double time_ms = (t2 - t1).seconds() * 1000;
+      double total_time_ms = (t2 - t0).seconds() * 1000;
 
       /* ---------- success temporary, check collision again ---------- */
+      // 收敛、达到最大迭代次数、已达到最小值或被停止，则进入碰撞检测阶段
       if (result == lbfgs::LBFGS_CONVERGENCE ||
           result == lbfgs::LBFGSERR_MAXIMUMITERATION ||
           result == lbfgs::LBFGS_ALREADY_MINIMIZED ||
           result == lbfgs::LBFGS_STOP)
       {
-        //ROS_WARN("Solver error in planning!, return = %s", lbfgs::lbfgs_strerror(result));
+        // ROS_WARN("Solver error in planning!, return = %s", lbfgs::lbfgs_strerror(result));
         flag_force_return = false;
 
         /*** collision check, phase 1 ***/
@@ -1485,7 +1598,7 @@ namespace ego_planner
           success = false;
           restart_nums++;
           initControlPoints(cps_.points, false);
-          new_lambda2_ *= 2;
+          new_lambda2_ *= 2; // 提高规避权重
 
           printf("\033[32miter(+1)=%d,time(ms)=%5.3f, swarm too close, keep optimizing\n\033[0m", iter_num_, time_ms);
 
@@ -1493,24 +1606,29 @@ namespace ego_planner
         }
 
         /*** collision check, phase 2 ***/
+        // 创建均匀的B样条曲线
         UniformBspline traj = UniformBspline(cps_.points, 3, bspline_interval_);
+        // 开始时间，结束时间
         double tm, tmp;
         traj.getTimeSpan(tm, tmp);
+        // 计算时间步长
         double t_step = (tmp - tm) / ((traj.evaluateDeBoorT(tmp) - traj.evaluateDeBoorT(tm)).norm() / grid_map_->getResolution());
+        // 遍历轨迹的前2/3部分进行障碍物检测
         for (double t = tm; t < tmp * 2 / 3; t += t_step) // Only check the closest 2/3 partition of the whole trajectory.
         {
           flag_occ = grid_map_->getInflateOccupancy(traj.evaluateDeBoorT(t));
           if (flag_occ)
           {
-            //cout << "hit_obs, t=" << t << " P=" << traj.evaluateDeBoorT(t).transpose() << endl;
+            // cout << "hit_obs, t=" << t << " P=" << traj.evaluateDeBoorT(t).transpose() << endl;
 
+            // 如果在前三个控制点范围内检测到了碰撞则视为不可行
             if (t <= bspline_interval_) // First 3 control points in obstacles!
             {
               // cout << cps_.points.col(1).transpose() << "\n"
               //      << cps_.points.col(2).transpose() << "\n"
               //      << cps_.points.col(3).transpose() << "\n"
               //      << cps_.points.col(4).transpose() << endl;
-              ROS_WARN("First 3 control points in obstacles! return false, t=%f", t);
+              RCLCPP_WARN(rclcpp::get_logger("rebound_optimize"), "First 3 control points in obstacles! return false, t=%f", t);
               return false;
             }
 
@@ -1521,7 +1639,7 @@ namespace ego_planner
         // cout << "XXXXXX" << ((cps_.points.col(cps_.points.cols()-1) + 4*cps_.points.col(cps_.points.cols()-2) + cps_.points.col(cps_.points.cols()-3))/6 - local_target_pt_).norm() << endl;
 
         /*** collision check, phase 3 ***/
-//#define USE_SECOND_CLEARENCE_CHECK
+// #define USE_SECOND_CLEARENCE_CHECK
 #ifdef USE_SECOND_CLEARENCE_CHECK
         bool flag_cls_xyp, flag_cls_xyn, flag_cls_zp, flag_cls_zn;
         Eigen::Vector3d start_end_vec = traj.evaluateDeBoorT(tmp) - traj.evaluateDeBoorT(tm);
@@ -1566,11 +1684,13 @@ namespace ego_planner
           flag_occ = true;
 #endif
 
+        // 如果没有检测到碰撞视为优化成功
         if (!flag_occ)
         {
           printf("\033[32miter(+1)=%d,time(ms)=%5.3f,total_t(ms)=%5.3f,cost=%5.3f\n\033[0m", iter_num_, time_ms, total_time_ms, final_cost);
           success = true;
         }
+        // 如果有碰撞则重新初始化
         else // restart
         {
           restart_nums++;
@@ -1580,6 +1700,7 @@ namespace ego_planner
           printf("\033[32miter(+1)=%d,time(ms)=%5.3f, collided, keep optimizing\n\033[0m", iter_num_, time_ms);
         }
       }
+      // 如果优化被强制取消
       else if (result == lbfgs::LBFGSERR_CANCELED)
       {
         flag_force_return = true;
@@ -1588,8 +1709,9 @@ namespace ego_planner
       }
       else
       {
-        ROS_WARN("Solver error. Return = %d, %s. Skip this planning.", result, lbfgs::lbfgs_strerror(result));
-        // while (ros::ok());
+        RCLCPP_WARN(rclcpp::get_logger("rebound_optimize"), 
+                                        "Solver error. Return = %d, %s. Skip this planning.", result, lbfgs::lbfgs_strerror(result));
+        // while (rclcpp::ok());
       }
 
     } while (
@@ -1599,6 +1721,7 @@ namespace ego_planner
     return success;
   }
 
+  // 使用L-BFGS方法对目标函数进行优化，得到重新分配时间后，光滑、拟合较好、动力学可行的轨迹。
   bool BsplineOptimizer::refine_optimize()
   {
     iter_num_ = 0;
@@ -1628,13 +1751,15 @@ namespace ego_planner
           result == lbfgs::LBFGS_ALREADY_MINIMIZED ||
           result == lbfgs::LBFGS_STOP)
       {
-        //pass
+        // pass
       }
       else
       {
-        ROS_ERROR("Solver error in refining!, return = %d, %s", result, lbfgs::lbfgs_strerror(result));
+        RCLCPP_ERROR(rclcpp::get_logger("refine_optimize"), 
+                                        "Solver error in refining!, return = %d, %s", result, lbfgs::lbfgs_strerror(result));
       }
 
+      // 使用优化后的控制点创建新的轨迹
       UniformBspline traj = UniformBspline(cps_.points, 3, bspline_interval_);
       double tm, tmp;
       traj.getTimeSpan(tm, tmp);
@@ -1645,6 +1770,7 @@ namespace ego_planner
         {
           // cout << "Refined traj hit_obs, t=" << t << " P=" << traj.evaluateDeBoorT(t).transpose() << endl;
 
+          // 将ref_pts存储为矩阵形式
           Eigen::MatrixXd ref_pts(ref_pts_.size(), 3);
           for (size_t i = 0; i < ref_pts_.size(); i++)
           {
@@ -1656,6 +1782,7 @@ namespace ego_planner
         }
       }
 
+      // 如果存在碰撞则调整参数并重新迭代
       if (!flag_safe)
         lambda4_ *= 2;
 
@@ -1664,11 +1791,12 @@ namespace ego_planner
 
     lambda4_ = origin_lambda4;
 
-    //cout << "iter_num_=" << iter_num_ << endl;
+    // cout << "iter_num_=" << iter_num_ << endl;
 
     return flag_safe;
   }
 
+  // 计算损失
   void BsplineOptimizer::combineCostRebound(const double *x, double *grad, double &f_combine, const int n)
   {
     // cout << "drone_id_=" << drone_id_ << endl;
@@ -1696,14 +1824,15 @@ namespace ego_planner
     calcTerminalCost(cps_.points, f_terminal, g_terminal);
 
     f_combine = lambda1_ * f_smoothness + new_lambda2_ * f_distance + lambda3_ * f_feasibility + new_lambda2_ * f_swarm + lambda2_ * f_terminal;
-    //f_combine = lambda1_ * f_smoothness + new_lambda2_ * f_distance + lambda3_ * f_feasibility + new_lambda2_ * f_mov_objs;
-    //printf("origin %f %f %f %f\n", f_smoothness, f_distance, f_feasibility, f_combine);
+    // f_combine = lambda1_ * f_smoothness + new_lambda2_ * f_distance + lambda3_ * f_feasibility + new_lambda2_ * f_mov_objs;
+    // printf("origin %f %f %f %f\n", f_smoothness, f_distance, f_feasibility, f_combine);
 
     Eigen::MatrixXd grad_3D = lambda1_ * g_smoothness + new_lambda2_ * g_distance + lambda3_ * g_feasibility + new_lambda2_ * g_swarm + lambda2_ * g_terminal;
-    //Eigen::MatrixXd grad_3D = lambda1_ * g_smoothness + new_lambda2_ * g_distance + lambda3_ * g_feasibility + new_lambda2_ * g_mov_objs;
+    // Eigen::MatrixXd grad_3D = lambda1_ * g_smoothness + new_lambda2_ * g_distance + lambda3_ * g_feasibility + new_lambda2_ * g_mov_objs;
     memcpy(grad, grad_3D.data() + 3 * order_, n * sizeof(grad[0]));
   }
 
+  // 计算优化后的损失
   void BsplineOptimizer::combineCostRefine(const double *x, double *grad, double &f_combine, const int n)
   {
 
@@ -1716,7 +1845,7 @@ namespace ego_planner
     Eigen::MatrixXd g_fitness = Eigen::MatrixXd::Zero(3, cps_.points.cols());
     Eigen::MatrixXd g_feasibility = Eigen::MatrixXd::Zero(3, cps_.points.cols());
 
-    //time_satrt = ros::Time::now();
+    // time_satrt = rclcpp::Clock().now();
 
     calcSmoothnessCost(cps_.points, f_smoothness, g_smoothness);
     calcFitnessCost(cps_.points, f_fitness, g_fitness);
