@@ -1,5 +1,4 @@
 #include "bspline_opt/uniform_bspline.h"
-#include <ros/ros.h>
 
 namespace ego_planner
 {
@@ -208,6 +207,7 @@ namespace ego_planner
 
   // void UniformBspline::recomputeInit() {}
 
+  // 将一组点转换为控制点
   void UniformBspline::parameterizeToBspline(const double &ts, const vector<Eigen::Vector3d> &point_set,
                                              const vector<Eigen::Vector3d> &start_end_derivative,
                                              Eigen::MatrixXd &ctrl_pts)
