@@ -3,7 +3,7 @@
 
 #include <Eigen/Eigen>
 #include <vector>
-#include <ros/ros.h>
+#include <rclcpp/rclcpp.hpp>
 
 #include <bspline_opt/uniform_bspline.h>
 #include <traj_utils/polynomial_traj.h>
@@ -21,7 +21,7 @@ namespace ego_planner
     vector<UniformBspline> local_traj_;
 
     double global_duration_;
-    ros::Time global_start_time_;
+    rclcpp::Time global_start_time_;
     double local_start_time_, local_end_time_;
     double time_increase_;
     double last_time_inc_;
@@ -33,7 +33,7 @@ namespace ego_planner
 
     bool localTrajReachTarget() { return fabs(local_end_time_ - global_duration_) < 0.1; }
 
-    void setGlobalTraj(const PolynomialTraj &traj, const ros::Time &time)
+    void setGlobalTraj(const PolynomialTraj &traj, const rclcpp::Time &time)
     {
       global_traj_ = traj;
       global_traj_.init();
@@ -210,7 +210,7 @@ namespace ego_planner
 
     int traj_id_;
     double duration_;
-    ros::Time start_time_;
+    rclcpp::Time start_time_;
     Eigen::Vector3d start_pos_;
     UniformBspline position_traj_, velocity_traj_, acceleration_traj_;
   };
@@ -221,7 +221,7 @@ namespace ego_planner
 
     int drone_id;
     double duration_;
-    ros::Time start_time_;
+    rclcpp::Time start_time_;
     Eigen::Vector3d start_pos_;
     UniformBspline position_traj_;
   };
