@@ -2,29 +2,20 @@
 
 using std::cout;
 using std::endl;
+
 namespace ego_planner
 {
-  PlanningVisualization::PlanningVisualization(ros::NodeHandle &nh)
-  {
-    node = nh;
-
-    goal_point_pub = nh.advertise<visualization_msgs::Marker>("goal_point", 2);
-    global_list_pub = nh.advertise<visualization_msgs::Marker>("global_list", 2);
-    init_list_pub = nh.advertise<visualization_msgs::Marker>("init_list", 2);
-    optimal_list_pub = nh.advertise<visualization_msgs::Marker>("optimal_list", 2);
-    a_star_list_pub = nh.advertise<visualization_msgs::Marker>("a_star_list", 20);
-  }
 
   // // real ids used: {id, id+1000}
-  void PlanningVisualization::displayMarkerList(ros::Publisher &pub, const vector<Eigen::Vector3d> &list, double scale,
-                                                Eigen::Vector4d color, int id, bool show_sphere /* = true */ )
+  void PlanningVisualization::displayMarkerList(rclcpp::Publisher<visualization_msgs::msg::Marker>::SharedPtr &pub, const vector<Eigen::Vector3d> &list, double scale,
+                                                Eigen::Vector4d color, int id, bool show_sphere /* = true */)
   {
-    visualization_msgs::Marker sphere, line_strip;
+    visualization_msgs::msg::Marker sphere, line_strip;
     sphere.header.frame_id = line_strip.header.frame_id = "world";
-    sphere.header.stamp = line_strip.header.stamp = ros::Time::now();
-    sphere.type = visualization_msgs::Marker::SPHERE_LIST;
-    line_strip.type = visualization_msgs::Marker::LINE_STRIP;
-    sphere.action = line_strip.action = visualization_msgs::Marker::ADD;
+    sphere.header.stamp = line_strip.header.stamp = rclcpp::Clock().now();
+    sphere.type = visualization_msgs::msg::Marker::SPHERE_LIST;
+    line_strip.type = visualization_msgs::msg::Marker::LINE_STRIP;
+    sphere.action = line_strip.action = visualization_msgs::msg::Marker::ADD;
     sphere.id = id;
     line_strip.id = id + 1000;
 
@@ -37,29 +28,29 @@ namespace ego_planner
     sphere.scale.y = scale;
     sphere.scale.z = scale;
     line_strip.scale.x = scale / 2;
-    geometry_msgs::Point pt;
+    geometry_msgs::msg::Point pt;
     for (int i = 0; i < int(list.size()); i++)
     {
       pt.x = list[i](0);
       pt.y = list[i](1);
       pt.z = list[i](2);
-      //if (show_sphere) sphere.points.push_back(pt);
+      // if (show_sphere) sphere.points.push_back(pt);
       line_strip.points.push_back(pt);
     }
-    //if (show_sphere) pub.publish(sphere);
-    pub.publish(line_strip);
+    // if (show_sphere) pub.publish(sphere);
+    pub->publish(line_strip);
   }
 
   // real ids used: {id, id+1}
-  void PlanningVisualization::generatePathDisplayArray(visualization_msgs::MarkerArray &array,
+  void PlanningVisualization::generatePathDisplayArray(visualization_msgs::msg::MarkerArray &array,
                                                        const vector<Eigen::Vector3d> &list, double scale, Eigen::Vector4d color, int id)
   {
-    visualization_msgs::Marker sphere, line_strip;
+    visualization_msgs::msg::Marker sphere, line_strip;
     sphere.header.frame_id = line_strip.header.frame_id = "map";
-    sphere.header.stamp = line_strip.header.stamp = ros::Time::now();
-    sphere.type = visualization_msgs::Marker::SPHERE_LIST;
-    line_strip.type = visualization_msgs::Marker::LINE_STRIP;
-    sphere.action = line_strip.action = visualization_msgs::Marker::ADD;
+    sphere.header.stamp = line_strip.header.stamp = rclcpp::Clock().now();
+    sphere.type = visualization_msgs::msg::Marker::SPHERE_LIST;
+    line_strip.type = visualization_msgs::msg::Marker::LINE_STRIP;
+    sphere.action = line_strip.action = visualization_msgs::msg::Marker::ADD;
     sphere.id = id;
     line_strip.id = id + 1;
 
@@ -72,7 +63,7 @@ namespace ego_planner
     sphere.scale.y = scale;
     sphere.scale.z = scale;
     line_strip.scale.x = scale / 3;
-    geometry_msgs::Point pt;
+    geometry_msgs::msg::Point pt;
     for (int i = 0; i < int(list.size()); i++)
     {
       pt.x = list[i](0);
@@ -86,14 +77,14 @@ namespace ego_planner
   }
 
   // real ids used: {1000*id ~ (arrow nums)+1000*id}
-  void PlanningVisualization::generateArrowDisplayArray(visualization_msgs::MarkerArray &array,
+  void PlanningVisualization::generateArrowDisplayArray(visualization_msgs::msg::MarkerArray &array,
                                                         const vector<Eigen::Vector3d> &list, double scale, Eigen::Vector4d color, int id)
   {
-    visualization_msgs::Marker arrow;
+    visualization_msgs::msg::Marker arrow;
     arrow.header.frame_id = "map";
-    arrow.header.stamp = ros::Time::now();
-    arrow.type = visualization_msgs::Marker::ARROW;
-    arrow.action = visualization_msgs::Marker::ADD;
+    arrow.header.stamp = rclcpp::Clock().now();
+    arrow.type = visualization_msgs::msg::Marker::ARROW;
+    arrow.action = visualization_msgs::msg::Marker::ADD;
 
     // geometry_msgs::Point start, end;
     // arrow.points
@@ -106,7 +97,7 @@ namespace ego_planner
     arrow.scale.y = 2 * scale;
     arrow.scale.z = 2 * scale;
 
-    geometry_msgs::Point start, end;
+    geometry_msgs::msg::Point start, end;
     for (int i = 0; i < int(list.size() / 2); i++)
     {
       // arrow.color.r = color(0) / (1+i);
@@ -130,11 +121,11 @@ namespace ego_planner
 
   void PlanningVisualization::displayGoalPoint(Eigen::Vector3d goal_point, Eigen::Vector4d color, const double scale, int id)
   {
-    visualization_msgs::Marker sphere;
+    visualization_msgs::msg::Marker sphere;
     sphere.header.frame_id = "world";
-    sphere.header.stamp = ros::Time::now();
-    sphere.type = visualization_msgs::Marker::SPHERE;
-    sphere.action = visualization_msgs::Marker::ADD;
+    sphere.header.stamp = rclcpp::Clock().now();
+    sphere.type = visualization_msgs::msg::Marker::SPHERE;
+    sphere.action = visualization_msgs::msg::Marker::ADD;
     sphere.id = id;
 
     sphere.pose.orientation.w = 1.0;
@@ -149,13 +140,13 @@ namespace ego_planner
     sphere.pose.position.y = goal_point(1);
     sphere.pose.position.z = goal_point(2);
 
-    goal_point_pub.publish(sphere);
+    goal_point_pub->publish(sphere);
   }
 
   void PlanningVisualization::displayGlobalPathList(vector<Eigen::Vector3d> init_pts, const double scale, int id)
   {
 
-    if (global_list_pub.getNumSubscribers() == 0)
+    if (global_list_pub->get_subscription_count() == 0)
     {
       return;
     }
@@ -167,36 +158,35 @@ namespace ego_planner
   void PlanningVisualization::displayMultiInitPathList(vector<vector<Eigen::Vector3d>> init_trajs, const double scale)
   {
 
-    if (init_list_pub.getNumSubscribers() == 0)
+    if (init_list_pub->get_subscription_count() == 0)
     {
       return;
     }
 
     static int last_nums = 0;
 
-    for ( int id=0; id<last_nums; id++ )
+    for (int id = 0; id < last_nums; id++)
     {
       Eigen::Vector4d color(0, 0, 0, 0);
       vector<Eigen::Vector3d> blank;
       displayMarkerList(init_list_pub, blank, scale, color, id, false);
-      ros::Duration(0.001).sleep();
+      rclcpp::sleep_for(std::chrono::milliseconds(1));
     }
     last_nums = 0;
 
-    for ( int id=0; id<init_trajs.size(); id++ )
+    for (int id = 0; id < init_trajs.size(); id++)
     {
       Eigen::Vector4d color(0, 0, 1, 0.7);
       displayMarkerList(init_list_pub, init_trajs[id], scale, color, id, false);
-      ros::Duration(0.001).sleep();
+      rclcpp::sleep_for(std::chrono::milliseconds(1));
       last_nums++;
     }
-
   }
 
   void PlanningVisualization::displayInitPathList(vector<Eigen::Vector3d> init_pts, const double scale, int id)
   {
 
-    if (init_list_pub.getNumSubscribers() == 0)
+    if (init_list_pub->get_subscription_count() == 0)
     {
       return;
     }
@@ -208,7 +198,7 @@ namespace ego_planner
   void PlanningVisualization::displayOptimalList(Eigen::MatrixXd optimal_pts, int id)
   {
 
-    if (optimal_list_pub.getNumSubscribers() == 0)
+    if (optimal_list_pub->get_subscription_count() == 0)
     {
       return;
     }
@@ -226,7 +216,7 @@ namespace ego_planner
   void PlanningVisualization::displayAStarList(std::vector<std::vector<Eigen::Vector3d>> a_star_paths, int id /* = Eigen::Vector4d(0.5,0.5,0,1)*/)
   {
 
-    if (a_star_list_pub.getNumSubscribers() == 0)
+    if (a_star_list_pub->get_subscription_count() == 0)
     {
       return;
     }
@@ -244,21 +234,21 @@ namespace ego_planner
       {
         list.push_back(pt);
       }
-      //Eigen::Vector4d color(0.5,0.5,0,1);
+      // Eigen::Vector4d color(0.5,0.5,0,1);
       displayMarkerList(a_star_list_pub, list, scale, color, id + i); // real ids used: [ id ~ id+a_star_paths.size() ]
       i++;
     }
   }
 
-  void PlanningVisualization::displayArrowList(ros::Publisher &pub, const vector<Eigen::Vector3d> &list, double scale, Eigen::Vector4d color, int id)
+  void PlanningVisualization::displayArrowList(rclcpp::Publisher<visualization_msgs::msg::MarkerArray>::SharedPtr &pub, const vector<Eigen::Vector3d> &list, double scale, Eigen::Vector4d color, int id)
   {
-    visualization_msgs::MarkerArray array;
+    visualization_msgs::msg::MarkerArray array;
     // clear
-    pub.publish(array);
+    pub->publish(array);
 
     generateArrowDisplayArray(array, list, scale, color, id);
 
-    pub.publish(array);
+    pub->publish(array);
   }
 
   // PlanningVisualization::
