@@ -6,7 +6,7 @@ namespace quadrotor_msgs
 {
 
 bool decodeOutputData(const std::vector<uint8_t> &data,
-                      quadrotor_msgs::OutputData &output)
+                      quadrotor_msgs::msg::OutputData &output)
 {
   struct OUTPUT_DATA output_data;
   if(data.size() != sizeof(output_data))
@@ -50,13 +50,13 @@ bool decodeOutputData(const std::vector<uint8_t> &data,
   //for(int i = 0; i < 4; i++)
   //  output.motor_rpm[i] = output_data.rpm[i];
 
-  output.seq = output_data.seq;
+  // output.seq = output_data.seq;
 
   return true;
 }
 
 bool decodeStatusData(const std::vector<uint8_t> &data,
-                      quadrotor_msgs::StatusData &status)
+                      quadrotor_msgs::msg::StatusData &status)
 {
   struct STATUS_DATA status_data;
   if(data.size() != sizeof(status_data))
@@ -65,13 +65,13 @@ bool decodeStatusData(const std::vector<uint8_t> &data,
 
   status.loop_rate = status_data.loop_rate;
   status.voltage = status_data.voltage/1e3;
-  status.seq = status_data.seq;
+  // status.seq = status_data.seq;
 
   return true;
 }
 
 bool decodePPROutputData(const std::vector<uint8_t> &data,
-                         quadrotor_msgs::PPROutputData &output)
+                         quadrotor_msgs::msg::PPROutputData &output)
 {
   struct PPR_OUTPUT_DATA output_data;
   if(data.size() != sizeof(output_data))
