@@ -1,10 +1,10 @@
 #ifndef __UAVUTILS_CONVERTERS_H
 #define __UAVUTILS_CONVERTERS_H
 
-#include <nav_msgs/Odometry.h>
-#include <geometry_msgs/Vector3.h>
-#include <geometry_msgs/Point.h>
-#include <geometry_msgs/Quaternion.h>
+#include <nav_msgs/msg/odometry.hpp>
+#include <geometry_msgs/msg/vector3.hpp>
+#include <geometry_msgs/msg/point.hpp>
+#include <geometry_msgs/msg/quaternion.hpp>
 
 #include <Eigen/Dense>
 #include <Eigen/Geometry>
@@ -12,8 +12,8 @@
 
 namespace uav_utils {
 
-inline void extract_odometry(nav_msgs::OdometryConstPtr msg, Eigen::Vector3d& p,
-                      Eigen::Vector3d& v, Eigen::Quaterniond& q)
+inline void extract_odometry(const nav_msgs::msg::Odometry::ConstPtr msg, Eigen::Vector3d& p, 
+                        Eigen::Vector3d& v, Eigen::Quaterniond& q)
 {
     p(0) = msg->pose.pose.position.x;
     p(1) = msg->pose.pose.position.y;
@@ -29,7 +29,7 @@ inline void extract_odometry(nav_msgs::OdometryConstPtr msg, Eigen::Vector3d& p,
     q.z() = msg->pose.pose.orientation.z;
 }
 
-inline void extract_odometry(nav_msgs::OdometryConstPtr msg, Eigen::Vector3d& p,
+inline void extract_odometry(const nav_msgs::msg::Odometry::ConstPtr msg, Eigen::Vector3d& p,
                       Eigen::Vector3d& v, Eigen::Quaterniond& q, Eigen::Vector3d& w)
 {
     extract_odometry(msg, p, v, q);
@@ -39,19 +39,18 @@ inline void extract_odometry(nav_msgs::OdometryConstPtr msg, Eigen::Vector3d& p,
     w(2) = msg->twist.twist.angular.z;
 }
 
-
 template <typename Scalar_t = double>
-Eigen::Matrix<Scalar_t, 3, 1> from_vector3_msg(const geometry_msgs::Vector3& msg) {
+Eigen::Matrix<Scalar_t, 3, 1> from_vector3_msg(const geometry_msgs::msg::Vector3& msg) {
     return Eigen::Matrix<Scalar_t, 3, 1>(msg.x, msg.y, msg.z);
 }
 
 template <typename Derived>
-geometry_msgs::Vector3 to_vector3_msg(const Eigen::DenseBase<Derived>& v) {
+geometry_msgs::msg::Vector3 to_vector3_msg(const Eigen::DenseBase<Derived>& v) {
 	EIGEN_STATIC_ASSERT_FIXED_SIZE(Derived);
     EIGEN_STATIC_ASSERT(Derived::RowsAtCompileTime == 3, THIS_METHOD_IS_ONLY_FOR_MATRICES_OF_A_SPECIFIC_SIZE);
     EIGEN_STATIC_ASSERT(Derived::ColsAtCompileTime == 1, THIS_METHOD_IS_ONLY_FOR_MATRICES_OF_A_SPECIFIC_SIZE);
 
-    geometry_msgs::Vector3 msg;
+    geometry_msgs::msg::Vector3 msg;
     msg.x = v.x();
     msg.y = v.y();
     msg.z = v.z();
@@ -59,17 +58,17 @@ geometry_msgs::Vector3 to_vector3_msg(const Eigen::DenseBase<Derived>& v) {
 }
 
 template <typename Scalar_t = double>
-Eigen::Matrix<Scalar_t, 3, 1> from_point_msg(const geometry_msgs::Point& msg) {
+Eigen::Matrix<Scalar_t, 3, 1> from_point_msg(const geometry_msgs::msg::Point& msg) {
     return Eigen::Matrix<Scalar_t, 3, 1>(msg.x, msg.y, msg.z);
 }
 
 template <typename Derived>
-geometry_msgs::Point to_point_msg(const Eigen::DenseBase<Derived>& v) {
+geometry_msgs::msg::Point to_point_msg(const Eigen::DenseBase<Derived>& v) {
 	EIGEN_STATIC_ASSERT_FIXED_SIZE(Derived);
     EIGEN_STATIC_ASSERT(Derived::RowsAtCompileTime == 3, THIS_METHOD_IS_ONLY_FOR_MATRICES_OF_A_SPECIFIC_SIZE);
     EIGEN_STATIC_ASSERT(Derived::ColsAtCompileTime == 1, THIS_METHOD_IS_ONLY_FOR_MATRICES_OF_A_SPECIFIC_SIZE);
     
-    geometry_msgs::Point msg;
+    geometry_msgs::msg::Point msg;
     msg.x = v.x();
     msg.y = v.y();
     msg.z = v.z();
@@ -77,19 +76,20 @@ geometry_msgs::Point to_point_msg(const Eigen::DenseBase<Derived>& v) {
 }
 
 template <typename Scalar_t = double>
-Eigen::Quaternion<Scalar_t> from_quaternion_msg(const geometry_msgs::Quaternion& msg) {
+Eigen::Quaternion<Scalar_t> from_quaternion_msg(const geometry_msgs::msg::Quaternion& msg) {
     return Eigen::Quaternion<Scalar_t>(msg.w, msg.x, msg.y, msg.z);
 }
 
 template <typename Scalar_t>
-geometry_msgs::Quaternion to_quaternion_msg(const Eigen::Quaternion<Scalar_t>& q) {
-    geometry_msgs::Quaternion msg;
+geometry_msgs::msg::Quaternion to_quaternion_msg(const Eigen::Quaternion<Scalar_t>& q) {
+    geometry_msgs::msg::Quaternion msg;
     msg.x = q.x();
     msg.y = q.y();
     msg.z = q.z();
     msg.w = q.w();
     return msg;
 }
+
 }
 
 #endif
