@@ -1,13 +1,13 @@
-#include <nav_msgs/Odometry.h>
-#include <nav_msgs/Path.h>
+#include <nav_msgs/msg/odometry.hpp>
+#include <nav_msgs/msg/path.hpp>
 #include <pcl/filters/voxel_grid.h>
 #include <pcl/kdtree/kdtree_flann.h>
 #include <pcl/point_cloud.h>
 #include <pcl/point_types.h>
 #include <pcl/search/kdtree.h>
 #include <pcl_conversions/pcl_conversions.h>
-#include <ros/ros.h>
-#include <sensor_msgs/PointCloud2.h>
+#include <rclcpp/rclcpp.hpp>
+#include <sensor_msgs/msg/point_cloud2.hpp>
 #include <Eigen/Dense>
 #include <fstream>
 #include <iostream>
@@ -17,21 +17,25 @@
 using namespace std;
 using namespace Eigen;
 
-ros::Publisher pub_cloud;
+// ROS2 初始化节点
+rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr pub_cloud;
 
-sensor_msgs::PointCloud2 local_map_pcl;
-sensor_msgs::PointCloud2 local_depth_pcl;
+sensor_msgs::msg::PointCloud2 local_map_pcl;
+sensor_msgs::msg::PointCloud2 local_depth_pcl;
 
-ros::Subscriber odom_sub;
-ros::Subscriber global_map_sub, local_map_sub;
 
-ros::Timer local_sensing_timer;
+rclcpp::Subscription<nav_msgs::msg::Odometry>::SharedPtr odom_sub;
+rclcpp::Subscription<sensor_msgs::msg::PointCloud2>::SharedPtr global_map_sub;
+rclcpp::Subscription<sensor_msgs::msg::PointCloud2>::SharedPtr local_map_sub;
 
-bool has_global_map(false);
-bool has_local_map(false);
-bool has_odom(false);
+rclcpp::TimerBase::SharedPtr local_sensing_timer;
 
-nav_msgs::Odometry _odom;
+
+bool has_global_map = false;
+bool has_local_map = false;
+bool has_odom = false;
+
+nav_msgs::msg::Odometry _odom;
 
 double sensing_horizon, sensing_rate, estimation_rate;
 double _x_size, _y_size, _z_size;
@@ -39,7 +43,7 @@ double _gl_xl, _gl_yl, _gl_zl;
 double _resolution, _inv_resolution;
 int _GLX_SIZE, _GLY_SIZE, _GLZ_SIZE;
 
-ros::Time last_odom_stamp = ros::TIME_MAX;
+rclcpp::Time last_odom_stamp = rclcpp::Time(0, 0, RCL_ROS_TIME);
 
 inline Eigen::Vector3d gridIndex2coord(const Eigen::Vector3i& index) {
   Eigen::Vector3d pt;
@@ -62,7 +66,8 @@ inline Eigen::Vector3i coord2gridIndex(const Eigen::Vector3d& pt) {
   return idx;
 };
 
-void rcvOdometryCallbck(const nav_msgs::Odometry& odom) {
+// 里程计信息回调
+void rcvOdometryCallbck(const nav_msgs::msg::Odometry& odom) {
   /*if(!has_global_map)
     return;*/
   has_odom = true;
@@ -71,23 +76,25 @@ void rcvOdometryCallbck(const nav_msgs::Odometry& odom) {
 
 pcl::PointCloud<pcl::PointXYZ> _cloud_all_map, _local_map;
 pcl::VoxelGrid<pcl::PointXYZ> _voxel_sampler;
-sensor_msgs::PointCloud2 _local_map_pcd;
+sensor_msgs::msg::PointCloud2 _local_map_pcd;
 
 pcl::search::KdTree<pcl::PointXYZ> _kdtreeLocalMap;
 vector<int> _pointIdxRadiusSearch;
 vector<float> _pointRadiusSquaredDistance;
 
-void rcvGlobalPointCloudCallBack(
-    const sensor_msgs::PointCloud2& pointcloud_map) {
+void rcvGlobalPointCloudCallBack(const sensor_msgs::msg::PointCloud2::SharedPtr pointcloud_map) {
   if (has_global_map) return;
 
-  ROS_WARN("Global Pointcloud received..");
+  RCLCPP_WARN(rclcpp::get_logger("rcvGlobalPointCloudCallBack"), "Global Pointcloud received..");
 
+  // 转换消息信息格式
   pcl::PointCloud<pcl::PointXYZ> cloud_input;
-  pcl::fromROSMsg(pointcloud_map, cloud_input);
+  pcl::fromROSMsg(*pointcloud_map, cloud_input);
 
+  // 使用体素滤波对点云降采样
   _voxel_sampler.setLeafSize(0.1f, 0.1f, 0.1f);
   _voxel_sampler.setInputCloud(cloud_input.makeShared());
+  // 结果保存在_cloud_all_map中
   _voxel_sampler.filter(_cloud_all_map);
 
   _kdtreeLocalMap.setInputCloud(_cloud_all_map.makeShared());
@@ -95,9 +102,10 @@ void rcvGlobalPointCloudCallBack(
   has_global_map = true;
 }
 
-void renderSensedPoints(const ros::TimerEvent& event) {
+void renderSensedPoints(/*const rclcpp::TimerBase event*/) {
   if (!has_global_map || !has_odom) return;
 
+  // 获取无人机姿态
   Eigen::Quaterniond q;
   q.x() = _odom.pose.pose.orientation.x;
   q.y() = _odom.pose.pose.orientation.y;
@@ -106,8 +114,10 @@ void renderSensedPoints(const ros::TimerEvent& event) {
 
   Eigen::Matrix3d rot;
   rot = q;
+  // 转换为旋转矩阵
   Eigen::Vector3d yaw_vec = rot.col(0);
 
+  // 清空并初始化点云数据
   _local_map.points.clear();
   pcl::PointXYZ searchPoint(_odom.pose.pose.position.x,
                             _odom.pose.pose.position.y,
@@ -116,24 +126,25 @@ void renderSensedPoints(const ros::TimerEvent& event) {
   _pointRadiusSquaredDistance.clear();
 
   pcl::PointXYZ pt;
+  // 进行半径搜索获取感知范围内的点
   if (_kdtreeLocalMap.radiusSearch(searchPoint, sensing_horizon,
                                    _pointIdxRadiusSearch,
                                    _pointRadiusSquaredDistance) > 0) {
+    // 遍历每个搜索到的点
     for (size_t i = 0; i < _pointIdxRadiusSearch.size(); ++i) {
       pt = _cloud_all_map.points[_pointIdxRadiusSearch[i]];
 
-      // if ((fabs(pt.z - _odom.pose.pose.position.z) / (pt.x - _odom.pose.pose.position.x)) >
-      //     tan(M_PI / 12.0))
-      //   continue;
+      // 设置最大仰角
       if ((fabs(pt.z - _odom.pose.pose.position.z) / sensing_horizon) >
           tan(M_PI / 6.0))
-        continue; 
+        continue;
 
-      Vector3d pt_vec(pt.x - _odom.pose.pose.position.x,
-                      pt.y - _odom.pose.pose.position.y,
-                      pt.z - _odom.pose.pose.position.z);
+      // 检查点是否在无人机视野内
+      Eigen::Vector3d pt_vec(pt.x - _odom.pose.pose.position.x,
+                             pt.y - _odom.pose.pose.position.y,
+                             pt.z - _odom.pose.pose.position.z);
 
-      if (pt_vec.normalized().dot(yaw_vec) < 0.5) continue; 
+      if (pt_vec.normalized().dot(yaw_vec) < 0.5) continue;
 
       _local_map.points.push_back(pt);
     }
@@ -141,6 +152,7 @@ void renderSensedPoints(const ros::TimerEvent& event) {
     return;
   }
 
+  // 设置局部地图属性并发布
   _local_map.width = _local_map.points.size();
   _local_map.height = 1;
   _local_map.is_dense = true;
@@ -148,39 +160,52 @@ void renderSensedPoints(const ros::TimerEvent& event) {
   pcl::toROSMsg(_local_map, _local_map_pcd);
   _local_map_pcd.header.frame_id = "map";
 
-  pub_cloud.publish(_local_map_pcd);
+  pub_cloud->publish(_local_map_pcd);
 }
 
 void rcvLocalPointCloudCallBack(
-    const sensor_msgs::PointCloud2& pointcloud_map) {
+    const sensor_msgs::msg::PointCloud2::SharedPtr pointcloud_map) {
   // do nothing, fix later
 }
 
 int main(int argc, char** argv) {
-  ros::init(argc, argv, "pcl_render");
-  ros::NodeHandle nh("~");
+  // 初始化ROS2
+  rclcpp::init(argc, argv);
 
-  nh.getParam("sensing_horizon", sensing_horizon);
-  nh.getParam("sensing_rate", sensing_rate);
-  nh.getParam("estimation_rate", estimation_rate);
+  // 创建节点
+  auto node = rclcpp::Node::make_shared("pcl_render");
 
-  nh.getParam("map/x_size", _x_size);
-  nh.getParam("map/y_size", _y_size);
-  nh.getParam("map/z_size", _z_size);
+  // 使用 declare_parameter 来声明参数并读取参数
+  node->declare_parameter("sensing_horizon", 0.0);
+  node->declare_parameter("sensing_rate", 0.0);
+  node->declare_parameter("estimation_rate", 0.0);
 
-  // subscribe point cloud
-  global_map_sub = nh.subscribe("global_map", 1, rcvGlobalPointCloudCallBack);
-  local_map_sub = nh.subscribe("local_map", 1, rcvLocalPointCloudCallBack);
-  odom_sub = nh.subscribe("odometry", 50, rcvOdometryCallbck);
+  node->declare_parameter("map/x_size", 0.0);
+  node->declare_parameter("map/y_size", 0.0);
+  node->declare_parameter("map/z_size", 0.0);
 
-  // publisher depth image and color image
-  pub_cloud =
-      nh.advertise<sensor_msgs::PointCloud2>("pcl_render_node/cloud", 10);
+  node->get_parameter("sensing_horizon", sensing_horizon);
+  node->get_parameter("sensing_rate", sensing_rate);
+  node->get_parameter("estimation_rate", estimation_rate);
+  node->get_parameter("map/x_size", _x_size);
+  node->get_parameter("map/y_size", _y_size);
+  node->get_parameter("map/z_size", _z_size);
 
+  // 订阅点云数据
+  global_map_sub = node->create_subscription<sensor_msgs::msg::PointCloud2>(
+      "global_map", 1, rcvGlobalPointCloudCallBack);
+  local_map_sub = node->create_subscription<sensor_msgs::msg::PointCloud2>(
+      "local_map", 1, rcvLocalPointCloudCallBack);
+  odom_sub = node->create_subscription<nav_msgs::msg::Odometry>(
+      "odometry", 50, rcvOdometryCallbck);
+
+  // 发布者：点云数据
+  pub_cloud = node->create_publisher<sensor_msgs::msg::PointCloud2>("pcl_render_node/cloud", 10);
+
+  // 定时器：控制渲染频率
   double sensing_duration = 1.0 / sensing_rate * 2.5;
-
-  local_sensing_timer =
-      nh.createTimer(ros::Duration(sensing_duration), renderSensedPoints);
+  local_sensing_timer = node->create_wall_timer(
+      std::chrono::duration<double>(sensing_duration), std::bind(&renderSensedPoints));
 
   _inv_resolution = 1.0 / _resolution;
 
@@ -188,15 +213,18 @@ int main(int argc, char** argv) {
   _gl_yl = -_y_size / 2.0;
   _gl_zl = 0.0;
 
-  _GLX_SIZE = (int)(_x_size * _inv_resolution);
-  _GLY_SIZE = (int)(_y_size * _inv_resolution);
-  _GLZ_SIZE = (int)(_z_size * _inv_resolution);
+  _GLX_SIZE = static_cast<int>(_x_size * _inv_resolution);
+  _GLY_SIZE = static_cast<int>(_y_size * _inv_resolution);
+  _GLZ_SIZE = static_cast<int>(_z_size * _inv_resolution);
 
-  ros::Rate rate(100);
-  bool status = ros::ok();
+  rclcpp::Rate rate(100);
+  bool status = rclcpp::ok();
   while (status) {
-    ros::spinOnce();
-    status = ros::ok();
+    rclcpp::spin_some(node);  
+    status = rclcpp::ok();
     rate.sleep();
   }
+
+  rclcpp::shutdown();
+  return 0;
 }
