@@ -1,19 +1,19 @@
 #include <pcl/point_cloud.h>
 #include <pcl/point_types.h>
-// #include <pcl/search/kdtree.h>
 #include <pcl/kdtree/kdtree_flann.h>
 #include <pcl_conversions/pcl_conversions.h>
-#include <iostream>
 
-#include <geometry_msgs/PoseStamped.h>
-#include <geometry_msgs/Vector3.h>
+#include <geometry_msgs/msg/pose_stamped.hpp>
+#include <geometry_msgs/msg/vector3.hpp>
+#include <nav_msgs/msg/odometry.hpp>
+#include <sensor_msgs/msg/point_cloud2.hpp>
+
 #include <math.h>
-#include <nav_msgs/Odometry.h>
-#include <ros/console.h>
-#include <ros/ros.h>
-#include <sensor_msgs/PointCloud2.h>
+#include <iostream>
 #include <Eigen/Eigen>
 #include <random>
+
+#include "rclcpp/rclcpp.hpp"
 
 using namespace std;
 
@@ -31,10 +31,11 @@ uniform_real_distribution<double> rand_w;
 uniform_real_distribution<double> rand_h;
 uniform_real_distribution<double> rand_inf;
 
-ros::Publisher _local_map_pub;
-ros::Publisher _all_map_pub;
-ros::Publisher click_map_pub_;
-ros::Subscriber _odom_sub;
+// 定义订阅者和发布者
+rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr _local_map_pub;
+rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr _all_map_pub;
+rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr click_map_pub_;
+rclcpp::Subscription<nav_msgs::msg::Odometry>::SharedPtr _odom_sub;
 
 vector<double> _state;
 
@@ -55,14 +56,15 @@ uniform_real_distribution<double> rand_radius2_;
 uniform_real_distribution<double> rand_theta_;
 uniform_real_distribution<double> rand_z_;
 
-sensor_msgs::PointCloud2 globalMap_pcd;
+sensor_msgs::msg::PointCloud2 globalMap_pcd;
 pcl::PointCloud<pcl::PointXYZ> cloudMap;
 
-sensor_msgs::PointCloud2 localMap_pcd;
+sensor_msgs::msg::PointCloud2 localMap_pcd;
 pcl::PointCloud<pcl::PointXYZ> clicked_cloud_;
 
+// 随机地图生成
 void RandomMapGenerate() {
-  pcl::PointXYZ pt_random;
+    pcl::PointXYZ pt_random;
 
   rand_x = uniform_real_distribution<double>(_x_l, _x_h);
   rand_y = uniform_real_distribution<double>(_y_l, _y_h);
@@ -147,15 +149,18 @@ void RandomMapGenerate() {
   cloudMap.height = 1;
   cloudMap.is_dense = true;
 
-  ROS_WARN("Finished generate random map ");
+  RCLCPP_WARN(rclcpp::get_logger("RandomMapGenerate"), "Finished generate random map"); 
 
   kdtreeLocalMap.setInputCloud(cloudMap.makeShared());
 
   _map_ok = true;
+
 }
 
+// 生成随机障碍，柱形和圆形
+// 相较于上面那个函数增加了距离限制和缩放因子
 void RandomMapGenerateCylinder() {
-  pcl::PointXYZ pt_random;
+    pcl::PointXYZ pt_random;
 
   vector<Eigen::Vector2d> obs_position;
 
@@ -171,7 +176,8 @@ void RandomMapGenerateCylinder() {
   rand_z_ = uniform_real_distribution<double>(z_l_, z_h_);
 
   // generate polar obs
-  for (int i = 0; i < _obs_num && ros::ok(); i++) {
+  // 生成柱形
+  for (int i = 0; i < _obs_num && rclcpp::ok(); i++) {
     double x, y, w, h, inf;
     x = rand_x(eng);
     y = rand_y(eng);
@@ -217,6 +223,7 @@ void RandomMapGenerateCylinder() {
   }
 
   // generate circle obs
+  // 生成圆形
   for (int i = 0; i < circle_num_; ++i) {
     double x, y, z;
     x = rand_x(eng);
@@ -264,173 +271,218 @@ void RandomMapGenerateCylinder() {
   cloudMap.height = 1;
   cloudMap.is_dense = true;
 
-  ROS_WARN("Finished generate random map ");
+  RCLCPP_WARN(rclcpp::get_logger("RandomMapGenerateCylinder"), "Finished generate random map "); 
 
+  // 将cloudmap转换为一个基于kd tree的点云地图
   kdtreeLocalMap.setInputCloud(cloudMap.makeShared());
 
   _map_ok = true;
 }
 
-void rcvOdometryCallbck(const nav_msgs::Odometry odom) {
-  if (odom.child_frame_id == "X" || odom.child_frame_id == "O") return;
-  _has_odom = true;
+// 里程计信息订阅回调
+void rcvOdometryCallback(const nav_msgs::msg::Odometry &odom) {
+    if (odom.child_frame_id == "X" || odom.child_frame_id == "O") return;
+    _has_odom = true;
 
-  _state = {odom.pose.pose.position.x,
-            odom.pose.pose.position.y,
-            odom.pose.pose.position.z,
-            odom.twist.twist.linear.x,
-            odom.twist.twist.linear.y,
-            odom.twist.twist.linear.z,
-            0.0,
-            0.0,
-            0.0};
+    _state = {
+        odom.pose.pose.position.x,
+        odom.pose.pose.position.y,
+        odom.pose.pose.position.z,
+        odom.twist.twist.linear.x,
+        odom.twist.twist.linear.y,
+        odom.twist.twist.linear.z,
+        0.0,
+        0.0,
+        0.0
+    };
 }
 
 int i = 0;
+// 发布点云信息
 void pubSensedPoints() {
-  // if (i < 10) {
-  pcl::toROSMsg(cloudMap, globalMap_pcd);
-  globalMap_pcd.header.frame_id = "world";
-  _all_map_pub.publish(globalMap_pcd);
-  // }
+    // 将点云转换为 ROS2 消息格式并发布
+    pcl::toROSMsg(cloudMap, globalMap_pcd);
+    globalMap_pcd.header.frame_id = "world";
+    _all_map_pub->publish(globalMap_pcd);
 
-  return;
+    return; // 有这个return后续的代码都不会执行
 
-  /* ---------- only publish points around current position ---------- */
-  if (!_map_ok || !_has_odom) return;
+    /* ---------- only publish points around current position ---------- */
+    // 只有地图生成完毕且有位置信息（里程计数据）时，才会发布局部地图
+    if (!_map_ok || !_has_odom) return;
 
-  pcl::PointCloud<pcl::PointXYZ> localMap;
+    pcl::PointCloud<pcl::PointXYZ> localMap;
 
-  pcl::PointXYZ searchPoint(_state[0], _state[1], _state[2]);
-  pointIdxRadiusSearch.clear();
-  pointRadiusSquaredDistance.clear();
+    // 设置搜索点
+    pcl::PointXYZ searchPoint(_state[0], _state[1], _state[2]);
+    pointIdxRadiusSearch.clear();
+    pointRadiusSquaredDistance.clear();
 
-  pcl::PointXYZ pt;
+    if (std::isnan(searchPoint.x) || std::isnan(searchPoint.y) || std::isnan(searchPoint.z))
+        return;
 
-  if (isnan(searchPoint.x) || isnan(searchPoint.y) || isnan(searchPoint.z))
-    return;
-
-  if (kdtreeLocalMap.radiusSearch(searchPoint, _sensing_range,
-                                  pointIdxRadiusSearch,
-                                  pointRadiusSquaredDistance) > 0) {
-    for (size_t i = 0; i < pointIdxRadiusSearch.size(); ++i) {
-      pt = cloudMap.points[pointIdxRadiusSearch[i]];
-      localMap.points.push_back(pt);
+    // 搜索感知范围内的点并构建局部地图
+    if (kdtreeLocalMap.radiusSearch(searchPoint, _sensing_range,
+                                    pointIdxRadiusSearch,
+                                    pointRadiusSquaredDistance) > 0) {
+        for (size_t i = 0; i < pointIdxRadiusSearch.size(); ++i) {
+            pcl::PointXYZ pt = cloudMap.points[pointIdxRadiusSearch[i]];
+            localMap.points.push_back(pt);
+        }
+    } else {
+        RCLCPP_ERROR(rclcpp::get_logger("map_server"), "No obstacles.");
+        return;
     }
-  } else {
-    ROS_ERROR("[Map server] No obstacles .");
-    return;
-  }
 
-  localMap.width = localMap.points.size();
-  localMap.height = 1;
-  localMap.is_dense = true;
+    // 发布局部地图
+    localMap.width = localMap.points.size();
+    localMap.height = 1;
+    localMap.is_dense = true;
 
-  pcl::toROSMsg(localMap, localMap_pcd);
-  localMap_pcd.header.frame_id = "world";
-  _local_map_pub.publish(localMap_pcd);
+    pcl::toROSMsg(localMap, localMap_pcd);
+    localMap_pcd.header.frame_id = "world";
+    _local_map_pub->publish(localMap_pcd);
 }
 
-void clickCallback(const geometry_msgs::PoseStamped& msg) {
-  double x = msg.pose.position.x;
-  double y = msg.pose.position.y;
-  double w = rand_w(eng);
-  double h;
-  pcl::PointXYZ pt_random;
+// 根据点击的位置，在地图中添加一个随机大小的柱状障碍物，并将其发布为一个局部地图
+void clickCallback(const geometry_msgs::msg::PoseStamped &msg) {
+    // 提取点的位置并生成障碍物
+    double x = msg.pose.position.x;
+    double y = msg.pose.position.y;
+    double w = rand_w(eng);
+    double h;
+    pcl::PointXYZ pt_random;
 
-  x = floor(x / _resolution) * _resolution + _resolution / 2.0;
-  y = floor(y / _resolution) * _resolution + _resolution / 2.0;
+    // 将点击的位置对齐到网格
+    x = std::floor(x / _resolution) * _resolution + _resolution / 2.0;
+    y = std::floor(y / _resolution) * _resolution + _resolution / 2.0;
 
-  int widNum = ceil(w / _resolution);
+    // 计算障碍物的宽度
+    int widNum = std::ceil(w / _resolution);
 
-  for (int r = -widNum / 2.0; r < widNum / 2.0; r++)
-    for (int s = -widNum / 2.0; s < widNum / 2.0; s++) {
-      h = rand_h(eng);
-      int heiNum = ceil(h / _resolution);
-      for (int t = -1; t < heiNum; t++) {
-        pt_random.x = x + (r + 0.5) * _resolution + 1e-2;
-        pt_random.y = y + (s + 0.5) * _resolution + 1e-2;
-        pt_random.z = (t + 0.5) * _resolution + 1e-2;
-        clicked_cloud_.points.push_back(pt_random);
-        cloudMap.points.push_back(pt_random);
-      }
+    // 生成障碍物
+    for (int r = -widNum / 2.0; r < widNum / 2.0; r++) {
+        for (int s = -widNum / 2.0; s < widNum / 2.0; s++) {
+            h = rand_h(eng);
+            int heiNum = std::ceil(h / _resolution);
+            for (int t = -1; t < heiNum; t++) {
+                pt_random.x = x + (r + 0.5) * _resolution + 1e-2;
+                pt_random.y = y + (s + 0.5) * _resolution + 1e-2;
+                pt_random.z = (t + 0.5) * _resolution + 1e-2;
+                clicked_cloud_.points.push_back(pt_random);
+                cloudMap.points.push_back(pt_random);
+            }
+        }
     }
-  clicked_cloud_.width = clicked_cloud_.points.size();
-  clicked_cloud_.height = 1;
-  clicked_cloud_.is_dense = true;
 
-  pcl::toROSMsg(clicked_cloud_, localMap_pcd);
-  localMap_pcd.header.frame_id = "world";
-  click_map_pub_.publish(localMap_pcd);
+    // 更新点云属性并发布局部地图
+    clicked_cloud_.width = clicked_cloud_.points.size();
+    clicked_cloud_.height = 1;
+    clicked_cloud_.is_dense = true;
 
-  cloudMap.width = cloudMap.points.size();
+    pcl::toROSMsg(clicked_cloud_, localMap_pcd);
+    localMap_pcd.header.frame_id = "world";
+    click_map_pub_->publish(localMap_pcd);
 
-  return;
+    cloudMap.width = cloudMap.points.size();
+
+    return;
 }
 
-int main(int argc, char** argv) {
-  ros::init(argc, argv, "random_map_sensing");
-  ros::NodeHandle n("~");
 
-  _local_map_pub = n.advertise<sensor_msgs::PointCloud2>("/map_generator/local_cloud", 1);
-  _all_map_pub = n.advertise<sensor_msgs::PointCloud2>("/map_generator/global_cloud", 1);
 
-  _odom_sub = n.subscribe("odometry", 50, rcvOdometryCallbck);
+int main(int argc, char **argv)
+{
+    rclcpp::init(argc, argv);
+    auto node = std::make_shared<rclcpp::Node>("random_map_sensing");
 
-  click_map_pub_ =
-      n.advertise<sensor_msgs::PointCloud2>("/pcl_render_node/local_map", 1);
-  // ros::Subscriber click_sub = n.subscribe("/goal", 10, clickCallback);
+    // 创建发布者
+    _local_map_pub = node->create_publisher<sensor_msgs::msg::PointCloud2>("/map_generator/local_cloud", 1);
+    _all_map_pub = node->create_publisher<sensor_msgs::msg::PointCloud2>("/map_generator/global_cloud", 1);
+    click_map_pub_ = node->create_publisher<sensor_msgs::msg::PointCloud2>("/pcl_render_node/local_map", 1);
 
-  n.param("init_state_x", _init_x, 0.0);
-  n.param("init_state_y", _init_y, 0.0);
+    // 创建订阅者
+    _odom_sub = node->create_subscription<nav_msgs::msg::Odometry>("odometry", 50, rcvOdometryCallback);
+    // auto click_sub = node->create_subscription<geometry_msgs::msg::PoseStamped>("/goal", 10, clickCallback);
 
-  n.param("map/x_size", _x_size, 50.0);
-  n.param("map/y_size", _y_size, 50.0);
-  n.param("map/z_size", _z_size, 5.0);
-  n.param("map/obs_num", _obs_num, 30);
-  n.param("map/resolution", _resolution, 0.1);
-  n.param("map/circle_num", circle_num_, 30);
+    // 声明和获取参数
+    node->declare_parameter("init_state_x", 0.0);
+    node->declare_parameter("init_state_y", 0.0);
+    node->declare_parameter("map/x_size", 50.0);
+    node->declare_parameter("map/y_size", 50.0);
+    node->declare_parameter("map/z_size", 5.0);
+    node->declare_parameter("map/obs_num", 30);
+    node->declare_parameter("map/resolution", 0.1);
+    node->declare_parameter("map/circle_num", 30);
 
-  n.param("ObstacleShape/lower_rad", _w_l, 0.3);
-  n.param("ObstacleShape/upper_rad", _w_h, 0.8);
-  n.param("ObstacleShape/lower_hei", _h_l, 3.0);
-  n.param("ObstacleShape/upper_hei", _h_h, 7.0);
+    node->declare_parameter("ObstacleShape/lower_rad", 0.3);
+    node->declare_parameter("ObstacleShape/upper_rad", 0.8);
+    node->declare_parameter("ObstacleShape/lower_hei", 3.0);
+    node->declare_parameter("ObstacleShape/upper_hei", 7.0);
 
-  n.param("ObstacleShape/radius_l", radius_l_, 7.0);
-  n.param("ObstacleShape/radius_h", radius_h_, 7.0);
-  n.param("ObstacleShape/z_l", z_l_, 7.0);
-  n.param("ObstacleShape/z_h", z_h_, 7.0);
-  n.param("ObstacleShape/theta", theta_, 7.0);
+    node->declare_parameter("ObstacleShape/radius_l", 7.0);
+    node->declare_parameter("ObstacleShape/radius_h", 7.0);
+    node->declare_parameter("ObstacleShape/z_l", 7.0);
+    node->declare_parameter("ObstacleShape/z_h", 7.0);
+    node->declare_parameter("ObstacleShape/theta", 7.0);
 
-  n.param("sensing/radius", _sensing_range, 10.0);
-  n.param("sensing/rate", _sense_rate, 10.0);
+    node->declare_parameter("sensing/radius", 10.0);
+    node->declare_parameter("sensing/rate", 10.0);
+    node->declare_parameter("min_distance", 1.0);
 
-  n.param("min_distance", _min_dist, 1.0);
+    node->get_parameter("init_state_x", _init_x);
+    node->get_parameter("init_state_y", _init_y);
+    node->get_parameter("map/x_size", _x_size);
+    node->get_parameter("map/y_size", _y_size);
+    node->get_parameter("map/z_size", _z_size);
+    node->get_parameter("map/obs_num", _obs_num);
+    node->get_parameter("map/resolution", _resolution);
+    node->get_parameter("map/circle_num", circle_num_);
 
-  _x_l = -_x_size / 2.0;
-  _x_h = +_x_size / 2.0;
+    node->get_parameter("ObstacleShape/lower_rad", _w_l);
+    node->get_parameter("ObstacleShape/upper_rad", _w_h);
+    node->get_parameter("ObstacleShape/lower_hei", _h_l);
+    node->get_parameter("ObstacleShape/upper_hei", _h_h);
 
-  _y_l = -_y_size / 2.0;
-  _y_h = +_y_size / 2.0;
+    node->get_parameter("ObstacleShape/radius_l", radius_l_);
+    node->get_parameter("ObstacleShape/radius_h", radius_h_);
+    node->get_parameter("ObstacleShape/z_l", z_l_);
+    node->get_parameter("ObstacleShape/z_h", z_h_);
+    node->get_parameter("ObstacleShape/theta", theta_);
 
-  _obs_num = min(_obs_num, (int)_x_size * 10);
-  _z_limit = _z_size;
+    node->get_parameter("sensing/radius", _sensing_range);
+    node->get_parameter("sensing/rate", _sense_rate);
+    node->get_parameter("min_distance", _min_dist);
 
-  ros::Duration(0.5).sleep();
+    // 地图边界和障碍物的设置
+    _x_l = -_x_size / 2.0;
+    _x_h = +_x_size / 2.0;
+    _y_l = -_y_size / 2.0;
+    _y_h = +_y_size / 2.0;
+    _obs_num = std::min(_obs_num, static_cast<int>(_x_size * 10));
+    _z_limit = _z_size;
 
-  unsigned int seed = rd();
-  // unsigned int seed = 2433201515;
-  cout << "seed=" << seed << endl;
-  eng.seed(seed);
+    rclcpp::sleep_for(std::chrono::milliseconds(500));
 
-  // RandomMapGenerate();
-  RandomMapGenerateCylinder();
+    // 初始化随机数生成器
+    unsigned int seed = rd();
+    // unsigned int seed = 2433201515;
+    std::cout << "seed=" << seed << std::endl;
+    eng.seed(seed);
 
-  ros::Rate loop_rate(_sense_rate);
+    // 生成随机地图
+    // RandomMapGenerate();
+    RandomMapGenerateCylinder();
 
-  while (ros::ok()) {
-    pubSensedPoints();
-    ros::spinOnce();
-    loop_rate.sleep();
-  }
+    // 设置循环频率并开始主循环
+    rclcpp::Rate loop_rate(_sense_rate);
+    while (rclcpp::ok()) {
+        // 发布感知到的点云数据
+        pubSensedPoints();
+        rclcpp::spin_some(node);
+        loop_rate.sleep();
+    }
+
+    rclcpp::shutdown();
+    return 0;
 }
