@@ -1,4 +1,5 @@
 #include <iostream>
+#include <deque>
 #include <pcl/point_cloud.h>
 #include <pcl/point_types.h>
 #include <pcl_conversions/pcl_conversions.h>
@@ -12,23 +13,22 @@
 #include <pcl/search/kdtree.h>
 #include <pcl_conversions/pcl_conversions.h>
 
-#include <geometry_msgs/PoseStamped.h>
-#include <geometry_msgs/Vector3.h>
-#include <nav_msgs/Odometry.h>
-#include <ros/console.h>
-#include <ros/ros.h>
-#include <sensor_msgs/PointCloud2.h>
+#include <geometry_msgs/msg/pose_stamped.hpp>
+#include <geometry_msgs/msg/vector3.hpp>
+#include <nav_msgs/msg/odometry.hpp>
+#include <rclcpp/rclcpp.hpp>
+#include <sensor_msgs/msg/point_cloud2.hpp>
 
 #include <Eigen/Eigen>
 #include <Eigen/SVD>
-#include <math.h>
+#include <cmath>
 
 #include <random>
 #include <sys/time.h>
 #include <time.h>
 
-#include "visualization_msgs/Marker.h"
-#include "visualization_msgs/MarkerArray.h"
+#include "visualization_msgs/msg/marker.hpp"
+#include "visualization_msgs/msg/marker_array.hpp"
 
 //! @todo historical above
 #include "maps.hpp"
@@ -46,17 +46,17 @@ pcl::search::KdTree<pcl::PointXYZ> kdtreeLocalMap;
 vector<int>                        pointIdxRadiusSearch;
 vector<float>                      pointRadiusSquaredDistance;
 
-ros::Publisher _local_map_pub;
-ros::Publisher _local_map_inflate_pub;
-ros::Publisher _global_map_pub;
+rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr _local_map_pub;
+rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr _local_map_inflate_pub;
+rclcpp::Publisher<sensor_msgs::msg::PointCloud2>::SharedPtr _global_map_pub;
 
-ros::Subscriber _map_sub;
-ros::Subscriber _odom_sub;
+rclcpp::Subscription<nav_msgs::msg::Odometry>::SharedPtr _map_sub;
+rclcpp::Subscription<nav_msgs::msg::Odometry>::SharedPtr _odom_sub;
 
-deque<nav_msgs::Odometry> _odom_queue;
-vector<double>            _state;
-const size_t              _odom_queue_size = 200;
-nav_msgs::Odometry        _odom;
+deque<nav_msgs::msg::Odometry> _odom_queue;
+vector<double>                 _state;
+const size_t                   _odom_queue_size = 200;
+nav_msgs::msg::Odometry        _odom;
 
 double z_limit;
 double _SenseRate;
@@ -66,22 +66,23 @@ double _sensing_range;
 bool map_ok    = false;
 bool _has_odom = false;
 
-sensor_msgs::PointCloud2       globalMap_pcd;
-sensor_msgs::PointCloud2       localMap_pcd;
+sensor_msgs::msg::PointCloud2 globalMap_pcd;
+sensor_msgs::msg::PointCloud2 localMap_pcd;
 pcl::PointCloud<pcl::PointXYZ> cloudMap;
-ros::Time                      begin_time = ros::TIME_MAX;
+rclcpp::Time begin_time = rclcpp::Time(0, 0);
 
 typedef Eigen::Vector3d ObsPos;
 typedef Eigen::Vector3d ObsSize; // x, y, height --- z
 typedef pair<ObsPos, ObsPos> Obstacle;
 std::vector<Obstacle> obstacle_list;
 
-void
-fixedMapGenerate()
+// 生成固定的障碍物地图
+void fixedMapGenerate()
 {
   double _resolution = 1.0;
 
   cloudMap.points.clear();
+  // 定义障碍物列表
   obstacle_list.push_back(
     make_pair(ObsPos(-7.0, 1.0, 0.0), ObsSize(1.0, 3.0, 5.0)));
   obstacle_list.push_back(
@@ -100,7 +101,8 @@ fixedMapGenerate()
   obstacle_list.push_back(
     make_pair(ObsPos(5.0, -2.5, 0.0), ObsSize(30.0, 1.0, 5.0)));
 
-  int           num_total_obs = obstacle_list.size();
+  // 遍历障碍物并将其体素化为点云
+  int num_total_obs = obstacle_list.size();
   pcl::PointXYZ pt_insert;
 
   for (int i = 0; i < num_total_obs; i++)
@@ -126,14 +128,14 @@ fixedMapGenerate()
     left_z  = 0;
     right_z = num_mesh_z;
 
+    // 将障碍物的边界体素添加到点云
     for (int r = left_x; r < right_x; r++)
       for (int s = left_y; s < right_y; s++)
       {
         for (int t = left_z; t < right_z; t++)
         {
           if ((r - left_x) * (r - right_x + 1) * (s - left_y) *
-                (s - right_y + 1) * (t - left_z) * (t - right_z + 1) ==
-              0)
+                (s - right_y + 1) * (t - left_z) * (t - right_z + 1) == 0)
           {
             pt_insert.x = x + r * _resolution;
             pt_insert.y = y + s * _resolution;
@@ -144,22 +146,23 @@ fixedMapGenerate()
       }
   }
 
+  // 设置点云属性并构建KD树
   cloudMap.width    = cloudMap.points.size();
   cloudMap.height   = 1;
   cloudMap.is_dense = true;
 
-  ROS_WARN("Finished generate random map ");
+  RCLCPP_WARN(rclcpp::get_logger("fixedMapGenerate"), "Finished generate random map ");
   cout << cloudMap.size() << endl;
   kdtreeLocalMap.setInputCloud(cloudMap.makeShared());
   map_ok = true;
 }
 
-void
-rcvOdometryCallbck(const nav_msgs::Odometry odom)
+// 里程计信息回调
+void rcvOdometryCallback(const nav_msgs::msg::Odometry::SharedPtr odom)
 {
-  if (odom.child_frame_id == "X" || odom.child_frame_id == "O")
+  if (odom->child_frame_id == "X" || odom->child_frame_id == "O")
     return;
-  _odom     = odom;
+  _odom     = *odom;
   _has_odom = true;
 
   _state = { _odom.pose.pose.position.x,
@@ -172,20 +175,19 @@ rcvOdometryCallbck(const nav_msgs::Odometry odom)
              0.0,
              0.0 };
 
-  _odom_queue.push_back(odom);
+  _odom_queue.push_back(*odom);
   while (_odom_queue.size() > _odom_queue_size)
     _odom_queue.pop_front();
 }
 
 int frequence_division_global = 40;
 
-void
-publishAllPoints()
+void publishAllPoints()
 {
   if (!map_ok)
     return;
 
-  if ((ros::Time::now() - begin_time).toSec() > 7.0)
+  if ((rclcpp::Clock().now() - begin_time).seconds() > 7.0)
     return;
 
   frequence_division_global--;
@@ -193,19 +195,16 @@ publishAllPoints()
   {
     pcl::toROSMsg(cloudMap, globalMap_pcd);
     globalMap_pcd.header.frame_id = kFrameIdNs_;
-    _global_map_pub.publish(globalMap_pcd);
+    _global_map_pub->publish(globalMap_pcd);
     frequence_division_global = 40;
-    ROS_ERROR("[SERVER]Publish one global map");
+    RCLCPP_ERROR(rclcpp::get_logger("publishAllPoints"), "[SERVER] Publish one global map");
   }
 }
 
-void
-pubSensedPoints()
+void pubSensedPoints()
 {
   if (!map_ok || !_has_odom)
     return;
-
-  // ros::Time time_bef_sensing = ros::Time::now();
 
   pcl::PointCloud<pcl::PointXYZ> localMap;
 
@@ -227,7 +226,7 @@ pubSensedPoints()
   }
   else
   {
-    // ROS_ERROR("[Map server] No obstacles .");
+    // RCLCPP_ERROR(rclcpp::get_logger("publishAllPoints"),"[Map server] No obstacles .");
     // cout<<searchPoint.x<<" , "<<searchPoint.y<<" , "<<searchPoint.z<<endl;
     // return;
   }
@@ -245,11 +244,11 @@ pubSensedPoints()
   pcl::toROSMsg(localMap, localMap_pcd);
 
   localMap_pcd.header.frame_id = kFrameIdNs_;
-  _local_map_pub.publish(localMap_pcd);
+  _local_map_pub->publish(localMap_pcd);
 
-  ros::Time time_aft_sensing = ros::Time::now();
+  rclcpp::Time time_aft_sensing = rclcpp::Clock().now();
 
-  if ((time_aft_sensing - begin_time).toSec() > 5.0)
+  if ((time_aft_sensing - begin_time).seconds() > 5.0)
     return;
 
   frequence_division_global--;
@@ -257,8 +256,8 @@ pubSensedPoints()
   {
     pcl::toROSMsg(cloudMap, globalMap_pcd);
     globalMap_pcd.header.frame_id = kFrameIdNs_;
-    _global_map_pub.publish(globalMap_pcd);
+    _global_map_pub->publish(globalMap_pcd);
     frequence_division_global = 40;
-    ROS_INFO("[SERVER]Publish one global map");
+    RCLCPP_INFO(rclcpp::get_logger("pubSensedPoints"), "[SERVER] Publish one global map");
   }
 }
